[roll] Roll fuchsia [scrutiny] Improve collection error logging.
Prior to this change failures in retrieving data collectors would return
a uuid that was very difficult to decode. This change provides a real
name to each collector and a short description of what is in the
collection. This greatly improves the ability to debug when a collection
cannot be processed.

Original-Bug: 81806
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564121
Original-Revision: ba71fd12f1ad4c56332f4032fdce6617cd3936af
GitOrigin-RevId: 4e84981fd22162b4a39e29a41089cd48a83d902a
Change-Id: Ib48dacd0aee8fd46874f447b13ec2b363fd31ead