[roll] Roll fuchsia [fs] Improve filesystem ID handling.
The Factoryfs filesystem ID was uninitialized. Hook this up.

Compute the blobfs legacy filesystem ID on demand. This is relatively
efficient and uncommon so doesn't warrant caching on the Blobfs object.

Store the zx::event in Minfs and Memfs. Previously these discarded the
event which negated the guarantee that the id is globally unique.

Some errors around creating and duplicating these events are now
ignored. The only time these will fail for non-misuse is if the kernel
is out of memory and in this case, prserving the exact error is unlikely
to be useful.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564123
Original-Revision: fa0b18c2b09f1e83399a05068e34ca538a094301
GitOrigin-RevId: 863f9c2bcca7d05a4c5f50b0814b41cf1ddbf83a
Change-Id: I9928d07677f815395633633ec1e58fe21484b52e