[roll] Roll fuchsia [assembly] Add options to modify update/base package names
Enable clients of assemble_system to specify the name of their update
and base packages. This is useful specifically for the "prime" images,
where we generate separate base and update packages and publish them to
the same repository as older base and update packages, but we do not
want the new packages to override the old.

Original-Bug: 79716
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/563168
Original-Revision: 2efbfb8f13c14c791ec0ad000d04996df97ca8b5
GitOrigin-RevId: 6e6180854942efb10f5bc360780e4ae7d5202e68
Change-Id: Ia404068ceaefa249a2a0d63821c04ee83cad3a12