[roll] Roll fuchsia [heap] Add virtual_alloc dependency and fix tests
This adds a, currently unused, build dependency between the heap and
the virtual_alloc library so that its tests are being executed.
Additionally this fixes a type coercion issue on the gcc builders that
was causing the tests to spuriously fail.

Original-Bug: 30973
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/563783
Original-Revision: b7f09065dff63a65f70106e7152e8ced7202980b
GitOrigin-RevId: a57b797716080f6c2b904aad6f9742d059410986
Change-Id: Ic1ada560d0bacf58717b8937084a768171299851