[roll] Roll fuchsia [wlan] Respect scan type in `wlan-dev connect`.
This change respects the `--scan-type` option in the `wlan-dev` tool's
`connect` command. After the removal of join scans in the SME,
`wlan-dev` temporarily performs these scans itself before attempting to
connect, but always performed an active scan regardless of the requested
scan type.

Today, this means that `wlan-dev connect` commands will not work if the
active scan type is not supported by the driver, as is currently the
case in `iwlwifi`.

Test: Manual testing and `wlantool-tests`.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/562145
Original-Revision: e37b73f46117a27037d5e41b9b12b7a958b25201
GitOrigin-RevId: 931a40e955533e0c903a91dfe7532a7fb303b9b8
Change-Id: Ib13aea31d6be68662c9a0969fc90fae73726c653