[roll] Roll fuchsia [scrutiny] Toolkit takes no dependency on the model.
The original intention of the toolkit plugin was to be a standalone thin
wrapper around Scrutiny utility code that allows the extraction of
blobfs and fvm. With the introduction of the packages toolkit option
that provides support for package extraction by urls this was no longer
possible. This change moves the package extractor into core and removes
the dependency from toolkit. Since toolkit is currently load bearing in
the verify_build python script this also prevents the Data Collector
being invoked in that script which is blocking landing early scheduler
failures.

Original-Bug: 81806
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564134
Original-Revision: 5a9d7164b7ee43a42d20574ab918ee0302717499
GitOrigin-RevId: 745d945fbaa8a198c1d8e82ba942d464a1297cca
Change-Id: I5d82bb1ebb3348c35c5b6f7e2aa4e289f5faaa12