[roll] Roll fuchsia [ffx][test] Improve message on error
Instead of returning a normal error, use ffx_bail when a test
fails.
When an error is returned noramlly, ffx prepends a 'BUG:' prefix. The
prefix is not added when ffx_bail is used. In most cases, when a test
fails this is not a bug in ffx or the test framework.

Testing: manual verification with `fx test` against a failing test.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/563391
Original-Revision: 428d44d36ce32606d9c90ecdddc569fedbd09df1
GitOrigin-RevId: 8831027bf00762c7c950e080cc2b2b7a2bdbe704
Change-Id: Id23e70e585e0aa8943789656e92b347333bb7902