[roll] Roll fuchsia [ffx] Improve speed of `ffx component run-legacy`
This plugin was using old `cs` code to traverse the hub and read
process IDs so that it could print them out when the component
was started.

Unfortunately, this functionality was not working properly because
the hub was read before the component was started.

Removing this functionality increases the speed of this command
dramatically.

As a result, `fx mem` can now use `ffx component run-legacy` again.

Original-Fixed: 81144
Test: fx test ffx_component_run_legacy
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564005
Original-Revision: 3c6bae204a7e90fc61ef941e1512ca86c4c323b6
GitOrigin-RevId: 801ebf3ce912081a6e73deea88a5a8e54cb8799f
Change-Id: I4a45de96fb4ecfc7eac0abc2be6e90d3bd931caf