[roll] Roll fuchsia [fxfs] Fix tombstone issue
Tombstone records didn't play nicely with directory records.  After
inserting a tombstone record, it's possible that directory child records
might exist in other layers, but the lookup function doesn't merge with
tombstone records, so it was possible to perform a lookup on a deleted
directory and find an old child.

To address this, we use the same key as object records but with a value
of None so that we will always know a deleted directory when we try to
open it.

In addition, we keep track of whether a directory is deleted within the
object-store layer, and return appropriate errors for all operations.

This addresses a flake which would manifest itself as an object
reference mismatch, since it could lead to updates on a deleted
directory which would resurrect the directory.

Multiply: fxfs-tests: 100
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564361
Original-Revision: 498daf78bfb72aee2ea00fdfd9bc967672ee6c89
GitOrigin-RevId: 04aa860f78899cf11f350604bca7b8e2cd30c95c
Change-Id: I473396b79f581bc2e6030a8f9dd0aa5a2e6bc989