[roll] Roll fuchsia [omaha-client] Policy return minimum wait for standard logic.
When using standard logic, even though we have accurate monotonic time,
it is possible that updates has been throttled, if a scheduled update
gets throttled, it will come back to ask `compute_next_update_time`
again without updating `last_update_time`, so in this case it could
return a time in the past or future, depends on the fuzzing, but
eventually it will always return a time in the past and results in an
infinite loop until the rate limit period is passed.

This CL fixes this case by adding minimum wait to the returned
`CheckTiming` even in standard logic.

Original-Fixed: 82023
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/563926
Original-Revision: a00bb5965fbbf9f9f454eafb98ec412c5933e4bc
GitOrigin-RevId: 311cd6210afa35f6207f5f7031ff60bcac87bd04
Change-Id: If77dafef1f843c0792eb5b79124ee9c9512ee0ba