[roll] Roll third_party/openthread [radio-spinel] fix incorrect RX channel after frame TX on another channel (#6837)
This commit fixes a bug in `RadioSpinel` platform implementation where
the request from OT core stack to enter RX mode may be ignored
causing the RCP/radio to stay in RX mode on an incorrect channel.

This situation can happen after a frame TX request on a different
channel (note that the `otRadioFrame` struct specifies the channel on
which the frame should be transmitted). After frame TX, radio driver
is expected to enter RX mode on the same channel on which the frame
TX happened. The OpenThread MAC layer always explicitly instructs the
radio to enter RX mode on the PAN channel (from `Mac::UpdateIdleMode()`
which in turn calls the radio platform API `otPlatRadioRecive()` with
the channel passed in as a parameter). The earlier `RadioSpinel`
platform implementation did not update `mChannel` on a tx request
which caused the subsequent `Receive()` to skip setting the channel on
RCP if the channel was same as before.

Original-Bug: 82094
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/third_party/openthread/+/564513
Original-Revision: 9546a44f0aac90466547989afdaf9428e4d5da21
GitOrigin-RevId: 3fb312b2be44ac8a922073458a55a4c84e9fcae4
Change-Id: I505c1612e20cca9fd47d6c13a2299551ca928fc2