[roll] Roll infra/recipes [specs] Clean up logic for loading spec files
Specs are now passed via properties instead of textproto files in
integration on all builders, so we no longer need this logic.

I replaced the "load spec" step with an "input properties" step where we
log *all* the input properties to the build, now that specs are just
another input property. That step technically generalizes to other
recipes, not just fuchsia.py, but it's most important for fuchsia.py
since it's the most heavily-used recipe, and it's purely a convenience,
as properties are always visible in the build's top-level "stderr" log.

Original-Bug: 77879
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/563112
Original-Revision: 532073de1401fb409dd5e9d8b782d17c0a0832bb
GitOrigin-RevId: 6cbf8f647390a970fb61a0db2c81a5d1d505b0ae
Change-Id: I9e55f0bae776abc0fef36d9d92911d2b8601274f