[roll] Roll fuchsia [fidl] Add anonymous layout usages to goldens
A new golden file containing anonymous layouts is added in
preparation for any special support that is added in the bindings
(i.e. so that behavior changes are observable in the goldens).

The feature gating of anonymous layouts is also removed to make
testing easier; This enables anonymous layouts before
fxbug.dev/74683 is fully finished but that's OK because
1. The remaining CLs (stacked above) will land shortly, and
2. The remaining changes won't break any users that decide
   to use the feature before 1

Test: fx regen-goldens
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/554981
Original-Revision: f121fd38588c901c31044dba5f1b1bd21244dcbe
GitOrigin-RevId: 92233926ffe00dab50580843d460f59789285aca
Change-Id: Ia1eb174d1089206c824de769248a9d169651a665