[roll] Roll fuchsia [network/tests] Log link local addresses on fail
Before this change, we see a rare flake that produces the error:

  Expected equality of these values:
    got_unknown_link_local_addrs
      Which is: 3
    want_unknown_link_local_addrs
      Which is: 4

Which is not actionable. Instead of counting the addresses, collect them
into a vector and assert on its size; on failure, the matcher will emit
the contents of the vector, producing a more actionable message.

Original-Bug: 82098
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/565123
Original-Revision: afeee2f18e5c32380ed4e708107d8794f745378e
GitOrigin-RevId: 08745daeb78e4cdc5602b9cbd6d4ac66ccdc4768
Change-Id: I388d8408231d4b8f94f0618aa6e4b1754aa7d69a