[roll] Roll fuchsia [scrutiny_config] Extract config into its own crate.
The scrutiny configuration crate should be independent of deps on
the framework. This allows for the flexibility of both the frontend
and the framework depending on the generic configuration. This is part
of a larger refactoring to improve build time configuration of Scrutiny
inside compiled actions.

Original-Bug: 81808
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564669
Original-Revision: 77a08724f52d0e60ffcdf9d6cd7b721be624a82a
GitOrigin-RevId: 7488acca182ce5aa10aaf86b5b4ddb504b550041
Change-Id: I5168d53be1bac3ea43db015bc632435863247a4e