[roll] Roll fuchsia [fidl][transformer] Fix bugs found by fuzzing the transformer
I manually ran the fuzzer on LLCPP + the transformer and it hit a small
number of assertions. This CL provides fixes so that it will return
cleanly and not hit an assertion at runtime.

In the future, I hope to run the fuzzer automatically on v2.

Original-Bug: 79177
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/565163
Original-Revision: 25b4094d4e90d46d48b32f3c9154c9dd4caf68a3
GitOrigin-RevId: 61faba56310e5e19a1e1dbd32ab90bcec29f78f5
Change-Id: I78b9f9cddf4f51d6268419f02f9df4e499aea778