[roll] Roll fuchsia [fxfs] Add a writeback cache
Introduces a writeback cache to fxfs which buffers writes on files until
they are fsynced() or their last handle closes.

This also changes the read path for files to read from the cache when
possible. Reading an absent part of the file will load the surrounding
32KiB of data into the cache for later use.

The writeback cache was written in a way where we can (hopefully) easily
move it out of fxfs once its design settles. Its only fxfs dependency is
on the allocator::Reservation construct, which we'll need to move out to
the VFS (with appropriate callbacks/hooks back to fxfs).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/547024
Original-Revision: a2937a20be215a766b1160712ea94d6d61d358f2
GitOrigin-RevId: 27ef41ddb1adc6a7cc495e60e7d4c9664861fce8
Change-Id: Ica3b7b7b9550e230c4d9a0cbfa3bb140d9517a2f