[roll] Roll fuchsia [sys][testing] TestWithEnvironmentFixture rename for soft migration
`TestWithEnvironment` will migrate to the SDK, without gTest dependencies.

This CL adds `TestWithEnvironmentFixture`, to replace `TestWithEnvironment`
usage, in- and out-of-tree, so existing out-of-tree dependencies on
`TestWithEnvironment` can be updated, first to `TestWithEnvironmentFixture`,
and later to the revised `TestWithEnvironment`.

This is the first migration step to resolving fxbug.dev/81468.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564570
Original-Revision: ba9bb4c7bb2700a3244e81e36381080c982f893f
GitOrigin-RevId: e75f3743f5de2ab89be64d4acb1bace25b0c090b
Change-Id: Idf24a703357feeb775718d739cd7500f9db5cd2d