[roll] Roll fuchsia [ffx][daemon] Move target et al into standalone lib.
This is the first (simplest) step for moving/refactoring target.rs as
well as making `Target` and `TargetCollection` visible in the services
library without making a circular dependency of daemon <-> service_lib.

This moves the following into its own crate:
* target.rs
* fastboot mod
* logger mod
* onet.rs
* ssh.rs
* zedboot.rs

The only notable changes are that many `pub(crate)` declarations were
changed to `pub` and `#[cfg(test)]` was removed from functions and
replaced with a doc comment stating the function is "intended for
testing." This was all done in lieu of a larger breakdown of
files/refactoring.

This also does a long-needed deps cleanup of the BUILD.gn files for target
and for the daemon.

Original-Bug: 81488
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/562223
Original-Revision: b75dc5b845fa06ba385a509a8e8b098f6461c9e3
GitOrigin-RevId: 9270c26431842d645d8c693eea83f8e67ee3ba9e
Change-Id: I8aae7a4cda2fab9996d8591e160c92df5efe2270