[roll] Roll fuchsia [fidl][transformer] Add more assertions about size bounds
This CL isn't in response to a particular issue - it was noticed that
primitive / enum / bits / handle / struct perform memcpy without a
immediately previous call to IncreaseNextOutOfLine. This means the
correctness of previous out of line block calculations is assumed and
the assertions added in this CL provide additional protection.

In addition, all assertions were changed to debug assertions because
that was the original intent.

Original-Bug: 79177
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/565121
Original-Revision: de757c18a7ded025ad337ceec4fe1eaea8db505b
GitOrigin-RevId: b715301ad26844be9b62dc1a8284cb8e2e0c34a5
Change-Id: I1ded7f2a7e60851b8b9f4d2761b2dc3267132345