[roll] Roll fuchsia [scrutiny] Remove ModelEnvironment in favor of ModelConfig.
Prior to this change the ModelEnvironment basically just copied all the values
from the ModelConfig. This was because the scrutiny_frontend crate dependend
on scrutiny_framework. Now scrutiny_config breaks this cycle so its possible
for scrutiny_framework to dependend on and use scrutiny_config. This
makes it simpler to add configuration for all collectors to use. This is
important in making Scrutiny more flexible in accepting variables from
the build in compiled actions.

Original-Bug: 81088
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564672
Original-Revision: 00c3b0f5d51a7109ffedc03fab522390ea1523d4
GitOrigin-RevId: 9d37c0f02f186694b34039d857272c4c97db7a0a
Change-Id: I9cb10cdd00564d319ce5ff94c540c3be6b9a40d5