[roll] Roll fuchsia [memfs] Add dispatcher to memfs.
In the future, the ManagedVfs will require a dispatcher. Convert memfs
to always require a dispatcher in the constructor.

Initialize some class members in the header for clarity.

Make GetMinfsBlksize to GetPageSize() and de-inline it (an inline
function with a static is suspicious).

Remove the name from the Memfs constructor (it was never used) and
convert the name in ::Create to a string_view (matches its eventual use
in a Dnode).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564135
Original-Revision: b055262602b89e91910720d6d2dcb1d4501468ea
GitOrigin-RevId: 01a7d745f8e33ade22af395d6a7c4741efde7677
Change-Id: Ia62230f7b47b1046f9729b751a7cf91031b416cf