[roll] Roll fuchsia [ffx][inspect] Fix hang from empty results
Requesting a snapshot from an ArchiveReader can return empty results if
for example you are asking for data for a non-existent component. The
iquery code explicitly sets this flag to not retry on an empty result. I
forgot to carry that over to the ffx plugin code that interacts with the
ArchiveReader.

Original-Fixed: 81981

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564322
Original-Revision: d1efce375458c1e82bea11a20301cf51a5eebbb6
GitOrigin-RevId: 4d03da88a64a80aa84a38352befed1a15c38a7ca
Change-Id: I423ee3bef9fed686802ac949db0be253c1f05682