[roll] Roll fuchsia [gigaboot] remove network spam
We had a few log messages that were printed whenever we saw a UDP packet
not meant for us, but this is a normal occurrence, commonly via unicast
flooding when a switch doesn't know which port a device is connected to
so has to send the packet to all of them.

Remove the logs, they're not helpful and get spammy if the user has a
network that triggers them often.

Original-Bug: 81988
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/565885
Original-Revision: 9bd2a26fecd8b27d584c601ad97fab2daa089194
GitOrigin-RevId: 311b256bd6a52f82968f051ae1b8764adccffd48
Change-Id: I148b21b0555b783e043624a1f86121965b03f8f3