[roll] Roll fuchsia [sdk] Add blobfs to the partner SDK
The image assembly tool (RFC-0072) requires the blobfs tool to construct
a BlobFS image. We are adding blobfs to the SDK so that the image
assembly tool is usable out of the tree.

We are naming the tool blobfs_do_not_depend to make it obvious that the
CLI surface is not expected to be used directly by anything outside of
the SDK, and is an implementation detail of the Image Assembler.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/541143
Original-Revision: 8ee331d917b5a699e67256ad4b4e94a765179da9
GitOrigin-RevId: fb6faf7010ae988511ff5d911c3849fbeb1f801f
Change-Id: I4809f59f363a5f4329c9b5c30baf6aa7be7f609b