[roll] Roll fuchsia [assembly] Add missing dep on blobfs image to flash.sh
In the refactor to using assemble_system(), the //:images build_api
target lost it's transitive dep on the blobfs image.  When the
temporary copy action was added, uefi_disk picked it up, but that's
not present on all builds, so flash.sh also needs a dep on it.

Original-Bug: 82274
Original-Bug: 195576132

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/565861
Original-Revision: dd917a9dc3674c54cd53f670f4142bd5996dbce9
GitOrigin-RevId: f6129953a54d379502d354c22c79620fd12d431c
Change-Id: I193d9f4467021b17e8569102f3ee52aad4cf8278