[roll] Roll fuchsia [mock_ddk] Convert max98373 driver tests to mock_ddk
This is part of a set of changes to discontinue use of the fake_ddk.
New drivers are instead recommended to use the mock_ddk or depend on
//src/devices/testing/no_ddk to get the necessary functions linked in.

Testing: This is a change to test code.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/549365
Original-Revision: a5a4f6c7dc4cd95e3b1cc2c49a3765f03c378c8d
GitOrigin-RevId: 00a0627f614ed6707119205c1362aa4157894fb6
Change-Id: I8ca82bccb3238efa191510cbbf9e5aeb9eb97a3a