[roll] Roll fuchsia [botanist][syslog] Re-add package repo if ssh disconnects.
If we lose SSH connection to the device, but are able to reconnect,
there's a possibility that the disruption was caused by a reboot which
could have removed the package repo config. This would show itself if
the syslog we start streaming in botanist gets interrupted but is able
to restart again (after the reboot). Thus, we can listen for errors from
the syslog streaming and re-add the package repo config so that it will
work for future tests in the shard.

Original-Bug: 81706
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/562301
Original-Revision: d67aac158e17ef3097afb577cbcfe54ac6bedd2c
GitOrigin-RevId: a0e2556b84a14f09b7bb465e52d313ac7f347de2
Change-Id: Ib083fe3329837ee6fdc05a307aa1d1decc573e1e