[roll] Roll fuchsia [realm-builder] Expose fuchsia.component.Binder from root
This CL changes the framework_intermediary library to expose
the protocol `fuchsia.component.Binder` from the root of a
constructed realm. This change was made to allow users to
start the realm in the event that have no other mechanism to
do so.

For more context on this migration, see
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0108_component_binder_protocol

Original-Bug: 81401
Original-Bug: 81400

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564562
Original-Revision: b56e3afaef082bcdb6490e265e3fd185e7328f11
GitOrigin-RevId: 673898daa6766ddca5c82b5ce57ff0ebde321edc
Change-Id: I90adea6c1468104bb33de6df6aed4ac081b4bedf