[roll] Roll fuchsia [netstack] Fix a flake in DHCP inspect tests
These tests send a packet to the client and then verify that various
stats have been incremented. When the sent packet arrives before the
DHCP client has begun listening for it, the packet is dropped and we
fail to observed the expected stats. Eliminate this race condition by
waiting for a message from the client before sending our packet in the
test runner, which ensures that the client will be ready by the time it
arrives.

Original-Bug: 80569
Multiply: netstack-inspect-integration-test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564702
Original-Revision: b21b841add844f63c6c2a7bec69c10a4181b6823
GitOrigin-RevId: f7f9cae818507fe512b06489808d848157aa965f
Change-Id: Id6c84326247d1ae6b581e1dfd7e2a77a7bc3183e