[roll] Roll fuchsia [starnix] Resolve race in FsNode::create_child
Previously, we could crash in FsNode::create_child if another thread was
dropping the child at the same time as we were trying to fetch it. The
other thread could have entered Drop::drop (having removed the last
refcount) and be waiting on the RwLock. The first thread would then see
a populated OnceCell with a Weak that failed to upgrade. Prior to this
CL, that upgrade failure would result in a panic.

This CL removes the OnceCell and we now use the Weak as the source of
truth for whether the entry in the cache is populated populated.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/566568
Original-Revision: 3c4230afae3407c804ca9ede3124d4f36e4c5301
GitOrigin-RevId: 3a3bf48759c6fc6bb6c5c35eb98e8aa928b4e520
Change-Id: I1a8dd42c5e19b0e78e1d9e8875b3b12c7622efcc