[roll] Roll fuchsia [component_manager] Fix panic test
This test was panicking because the configuration file specified
was incorrect. The intention of this test is to fail because the
URL of the root component could not be resolved.

Set a custom configuration that will be guaranteed to work with
this test.

Test: fx test component_manager_panic_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/565814
Original-Revision: f6fd9579d5ad48c8c412d610a6a0ca412ae19a44
GitOrigin-RevId: 3ad295c40b4167e6ca4c07d5ede04be4bb5a0d6a
Change-Id: I2b919126bc2bfcc4e71b7f1508e0b942049492e0