[roll] Roll fuchsia [vfs] Remove ManagedVfs default constructor.
The PagedVfs functionality will be moving into ManagedVfs and PagedVfs
requires a dispatcher in the constructor. This removes the default
constructor and updates the callers.

To tame some out-of-control ifdefs, a typedef is provided for the
cross-platform functions that need to take a dispatcher on Fuchsia and
nothing on the host.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564136
Original-Revision: 1ff7dacc68202e8aae0027abe352271308e987de
GitOrigin-RevId: 220e3933ebf2046e5d527414ce17e7de7588bb5e
Change-Id: Ibacbe5d3f1b6c1508310a2580d5d4ba7af37f8a4