[roll] Roll fuchsia [assembly] Remove vbmeta salt parameter
The transitional salt parameter in the VBMetaConfig is not longer used.
The salt parameter was used for sharing the same salt between two sets
of images, so that they could be binary equivalents. Now that we are not
comparing two sets of images, this is unused and can be removed.

Original-Bug: 77114
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564324
Original-Revision: c447e45d59ca79608dcb9eb94b6931c3bc50f506
GitOrigin-RevId: 78d0b752c497fac54a2a7ca7927f03569b8627b5
Change-Id: I5d31251e9be2e870eb29cea2ecfeb37c4d555af1