[roll] Roll fuchsia [archivist] Reduce log verbosity when handling an error directory ready event
Since we introduced the inspect shard, many test root CMLs had to transitively
include the shard. This led to some log spam which confuses people when
their test is failing for unrelated reasons.

An alternative to this implementation would be to not emit Error events
when a component "exposes" a directory to framework but never actually
serves it. I believe that the current (and tested) implementation in component
manager of emitting Error payloads when the component didn't serve the
directory but exposed it is correct and instead we can just reduce the
log verbosity here.

Original-Fixed: 82010

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/566809
Original-Revision: 7b3292942ad7c6a05f39476537084df01234fea1
GitOrigin-RevId: df50ed15613f2de6be55f1013015ae7de98cb8c9
Change-Id: I15e684e1e301a5e1cd54dffb4cc14497c29678d8