[roll] Roll fuchsia [driver_manager] Fix driver index fallback drivers
There's a bug for boot fallback drivers when they are loaded
in a device that requires base drivers.

The boot fallback drivers should not be loaded until the base
drivers are loaded. However, we are passing 'ignore_boot_drivers'
when we go to load all of the base drivers.

This flag is changed to 'return_only_base_and_fallback_drivers' because
we should be returning the newly found base drivers as well
as all of the boot fallback drivers that we haven't bound yet.

This CL also makes the isolated-devmgr-v2 'require-system' and
passes it an empty system.

TEST: fx test driver-manager-tests
      fx test ddk-fallback-test-v2

Original-Fixed: 82424

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/567122
Original-Revision: 3e8f53af20547ad498ab3df9e28b370bcbbd1641
GitOrigin-RevId: 22f3b17f551170ffb29222be8eb5573950728ef5
Change-Id: I95cedfd68b99735387638828d38dfbd3e56cbfd7