[roll] Roll fuchsia [hlcpp] V2 wire format encoding for HLCPP
V2 decoding support was previously added for HLCPP, but not encoding
support. This CL adds encoding support. There currently isn't a
validator that supports v2 validation so this CL runs the v1 validator
and disables tests that break due to differences (only 8).

It is not enabled in the live path and is only enabled for GIDL tests.

Original-Bug: 81995

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564061
Original-Revision: 9c931800bb73efd5b0329eb952c1aec6ac28dfef
GitOrigin-RevId: 2417010ca68c48235ba4d93bb3416f74da1e8e90
Change-Id: I6743e2ba7f3cf9cfed2fe118e2ed786eb3a05dc3