[roll] Roll fuchsia [kernel][x86] page_tables: Test for PTE global bit when queueing inval
When queueing invalidations set the global bit based on a PTE's global
bit, rather than on whether it is a kernel (vs user) address. The two
predicates come out the same way currently, but if there are uses of the
global bit for non-kernel addresses in the future, the current predicate
would be wrong.

Resolves an old TODO from teisenbe@.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/566567
Original-Revision: 5baa9c36440fa842b241b2831c2a2d34aa183f3a
GitOrigin-RevId: a066f9fbd9d96b31752be0ad85df010f9d1c7be5
Change-Id: Ib5f2e3bcfa00c3eda34eac3bfd18963adf46ed93