[roll] Roll fuchsia [settings] Wait for camera devices to init watcher
The initialization order is not guaranteed between the camera api and
the camera watcher agent, so there may be a valid empty response to
watching the devices. This change introduces a timeout during which the
agent will continue to watch for devices before deciding it can't find a
camera device.

Original-Fixed: 79768
Test: fx test setui_service_tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564139
Original-Revision: b577d8cef9da491f8fe5a279b3960ea797996519
GitOrigin-RevId: 389e385a9de3294d314fef9d0c790ec71bb6bcb0
Change-Id: I558b73121152afdbe5012be2cea1e5e54f581ed8