[roll] Roll fuchsia [fidlc] Always specify naming context
Previously, we would only specify the naming context in the IR for
anonymous layouts (since otherwise, it is implicitly just the
naming of the layout). This changes naming context to always
be present, renaming it from maybe_naming_context to naming_context.

This is mainly for clarity, so that things are explicit in the IR
and fidlgen does not need to make the assumption listed above.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/567301
Original-Revision: 243ae964a3a8dc09e195914340609d6f04a33507
GitOrigin-RevId: 902401c72bb5476e24a4230e8368dd64c5c2dd93
Change-Id: I3bfa4c0033e8627f3e8eb54e41b5d8f317ae9310