[roll] Roll experiences [e2e] Deflake 'screen is not black' test on ermine
The test is flaking because of a race condition when the screen is
rendered and the screenshot taken. The intention is to ensure that the
screen is rendered before the screenshot is taken. This change uses
flutter driver's screeshot method before taking scenic's screenshot to
increase the odds.

Original-Bug: 79921

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/567561
Original-Revision: 4ea7b085a9607ae6a165b8521c863aad385f0814
GitOrigin-RevId: d96ab5d50ea643431d2d6ed77ec3078ba3835e4c
Change-Id: Ib6eff830f3d24566bd0cf2bc47dc2dfcabca5fed