[roll] Roll fuchsia [input] Make the input pipeline fully asynchronous
See proposal at:
https://groups.google.com/a/fuchsia.dev/g/ui-input-dev/c/v3iGIYzMWR8

This change attaches the individual handlers in the input pipeline
together via async queues. This allows fully asynchronous input event
processing.

In the old code, while the top-level input of the pipeline was
asynchronous, the event processing was done in a for loop. This was
adequate for all pipeline stages that only generate events due to an
external stimulus, as the nested for loop would drain all the generated
events. However, for autorepeat generation this is not enough: we need
to inject key autorepeat events which get generated as result of a
continuous key actuation, and which are triggered by a timer.

Such a stage needs the handlers to behave as producer/consumer pairs,
so we refactored the pipeline to match. This allows us to implement the
autorepeat as a regular pipeline stage.

The correct initialization protocol for an input pipeline with
producer/consumer internal implementation is a bit more involved, but
this should not be relevant to any users as the change is purely
internal.

Some additional drive-by refactorings

- Made all fields on `InputPipeline` private.  This should make it
  easier to control how the fields are used.
- device_bindings was simplified.  Not sure if this was the intended
  semantics.

Note, the refactor didn't go all the way to making InputPipelineAssembly
a full-fledged builder.  This may be a next step before implementing
autorepeat.

Original-Fixed: 82352
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/566673
Original-Revision: 3280cf94c51b1f20d5738349fff4eeed4a8fd79f
GitOrigin-RevId: 142e1fe849b04ae43c47190952d512728d58659e
Change-Id: I4899dea618a2b29f1a12f024f499c21a5d333821