[roll] Roll fuchsia [fidl][cpp] NaturalClientBase and unified client wrappers
This CL adds |fidl::Client| and |fidl::SharedClient|, the unified
counterparts to |fidl::WireClient| and |fidl::WireSharedClient|. The
interface documentations are largely identical with the exception of the
`.wire()` method and some mentions of wire messages. No natural
messaging layer is generated yet; as a starting point we are testing
that the client wrappers themselves work.

Given a unified client:

  fidl::Client<SomeProtocol> client;

the wire methods are exposed under:

  client.wire()->Foo(...);

the natural methods are exposed under:

  client->Foo(...);

See the design doc linked in fxbug.dev/60240 for more details.

In order to expose two different set of methods, inheritance cannot be
used. We introduce |NaturalClientBase| as the base class for all future
natural client implementations, to be exposed by the `->` operator.
|NaturalClientBase| contains a |NaturalClientMessenger| which borrows
into the |ClientBase| that is part of the wire client implementation
object.

In the future, we could refactor WireClientImpl to also use a messenger
object, instead of inheriting from |ClientBase| and leaking a bunch of
internal methods to the user (fxbug.dev/82085).

Original-Bug: 60240
Test: fx test //src/lib/fidl/cpp
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/565681
Original-Revision: 808cd7d90a804262a71ded50c0ebf0ac5b50c16e
GitOrigin-RevId: 7f26e9b19c19683e1b69e3d87e3defdf5a931300
Change-Id: I8e08f6066fa753eecc9dcef5ab6b6653da79f0ab