[roll] Roll fuchsia [build] Introduce packaged_shared_library() template.
Add //build/dist/packaged_shared_library.gni which is essentially
a rewrite of variant_shared_library_redirect(), and deprecate the
latter (since it is still used by //vendor/... build rules).

The template is used to guarantee that a shared_library() instance
is always installed into a Fuchsia package, even if it is not an
actual ELF runtime dependency of one of the final executable or
loadable module targets in the package.

Original-Bug: 82489

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/567764
Original-Revision: 4cdb2da66fc3278e30c8f567c4e342c1cc462cde
GitOrigin-RevId: 51afd8474780e468c1062109fd4a7f5dd1a65739
Change-Id: I127f8360e57b3b0c24eb636fd87fd8d0b0c1da6b