[roll] Roll fuchsia [component_manager] increase timeout for test
The model::component::tests::stop_component_well_behaved_component_stop
test is flaking in CI. It flakes because it takes too long to stop the
component.

The timeouts for this are very short, 5 milliseconds for the stop
timeout, and are likely insufficient when the test bot is under
significant load. To resolve the flake, 10x the timeouts.

Original-Fixed: 82609
Multiply: component-manager-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/568121
Original-Revision: 7a2ec75049c23c5829518ec1427936a3bec5a646
GitOrigin-RevId: 6eed4e3f9c0f4c139e9ce53b6be7c05f6e9949bf
Change-Id: I53ff52f04617109663c0b11896f620e3b7734c30