[roll] Roll fuchsia [mold] Added a self-driving layer optimizer.
This patch adds a per-tile layer optimizer with passes that now
drives the tile painting and stores the layer state from one tile
to the next.

Driving the painting of a tile is now done through a new trait
called LayerPainter that abstracts over some the functionality
in Painter that is relevant to the driver. This simplifies the
structure of the Painter and strips it of any optimization checks
and layer bookkeeping.

That funcionality is now moved to the LayerWorkbench that now
treats coverage and segment data more abstractly, thus simplifying
algorithms that need information about all layers present in a tile.

The LayerWorkbench currently implements 3 optimization passes that
make extensive use of layer information:

  * a tile skipping pass that efficiently identifies whether the
    contents of the current tile have changed, emitting an op to
    skip the tile
  * a clip skipping pass that tries to identify when clipping can
    be completely skipped; namely:

    * when a layer is completely out of a clip, the layer gets
      skipped
    * when a tile is ompletely inside of a clip, the clip gets
      skipped and all layers skip applying needless masks
    * when a clip does not affect any subsequent layers, the clip
      gets skipped

  * an opacity pass that skips all layers underneath an opaque
    layer if that layer is complete and tries to emit an op to fill
    the tile with a solid color if all tiles above it are complete

In the description above 'skip' means that the Painter will be
completely unaware of the layer's existence, while 'complete' means
that the geometry will cover all the pixels of the tiles
completely.

The optimization passes make use of a MaskedVec type that can
efficiently soft-drop any items or skip first N of them. It also
enables these operations while it is iterated over.

Testing: fx test mold-tests && fx test surpass-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/568122
Original-Revision: ad4e5270105bf077076306385144e94c390f6863
GitOrigin-RevId: 51136f403777b6f438ba381588442d9d8d9b2507
Change-Id: I8c472d2eb848e0e43df1eccfc95ab9d4e7e31d75