[roll] Roll fuchsia [archivist] Refactor pipeline initialization for simplification
The archivist initialization code is a bit large and can be logically
subdivided. This tackles the pipeline initialization.

Additionally, this simplifies the work needed to create a new pipeline.
For example, if we tackle fxbug.dev/65801 and create a new pipeline for
sampler metrics, it now takes only a few lines of code to do so while
getting all the instrumentation and boilerplate for free.

Tested: archivist unit tests and integration tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/567530
Original-Revision: dc605d0918a8d0e54e35aa643bdf02bc6202b739
GitOrigin-RevId: 3d8248574690c13b9320fb451f7683c3404dd54b
Change-Id: Ifd1f812f67cd0ecf9fa580d0d3d6315eb1333e97