[roll] Roll fuchsia [libc] Add definition and stub for memfd_create()
memfd_create() is an extension to the sys/mman.h interface supported on
Linux, some BSDs, and Fuchsia. The Fuchsia implementation is implemented
inside fdio. This adds the prototype definition to the libc header along
with a weak stub so it can be used by programs in the normal way.

Currently no flags are supported.

Original-Fixed: 60236
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/568505
Original-Revision: fd1bb478b7634fd3b1840bb893fcdb7319e7fb31
GitOrigin-RevId: 62a6d445d04f4e244f8b910cac6db3f76def4410
Change-Id: I60bb9ad03975643cbb4c80df8d60c6eb4753b9f6