[roll] Roll fuchsia [scrutiny] Enable configurable plugin loading.
Prior to this changes plugins were not actually configurable through the
config file. This change allows plugins to register but not load for
certain use cases. This allows more fine grain control in apps as to which
specific plugins are running. It is up to the user of the API to correctly
set the plugins they need and all dependencies. This now allows for the
toolkit plugin to not use the core plugin at all.

Original-Bug: 81088
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/565897
Original-Revision: 18b2bc2f4f16a7c304df1fb4e6d922dfbb743b02
GitOrigin-RevId: 2d555e5537b7d2d618f69eba6f68bc6fe50a860a
Change-Id: I0d9994cd8dcf8ea754dd2bbbd790204f5cd142eb