[roll] Roll fuchsia [audio_core] Power-down if no active renderers
To reduce the idle power consumption of audio devices, use the
number of active audio renderers per-usage to power-down any audio
outputs (or output-channel subsets) that are unused.

In solving fxb/75709 and fxb/75714, we make these observations:
0) on many small consumer devices, the audio output amplifiers draw a
   significant fraction of the total power draw;
1) a significant fraction of the time, audio output is not needed
   (or only needed for proximity detection, if supported);
2) depending on hardware design, audio output amps can be powered-down
   and later powered-up, with few negative ramifications if any);
3) on some audio devices, ultrasound output requires only a subset of
   the audio output hardware, so any other outputs can be powered-down
   when no humanly-audible content (only ultrasound) is playing.

An IdlePolicy class uses ActiveStreamCountReporter to track changes
in the number of active "audible" and ultrasound renderers, by
RenderUsage, We overlay this with a current mapping from RenderUsage
to AudioDevice to determine the set of AudioDevices that are targets
of currently-active audible renderers, as well as the set of those
that are targets of currently-active ultrasonic renderers. Using
information from the driver that maps each output channel to a
frequency range, we use SetActiveChannels to inform the device when
certain channels no longer contain content that should be played; at
that point the device can safely power-down any hardware unrelated to
the remaining active channels.

For responsiveness reasons, we want to be slow in disabling hardware,
and fast in reenabling it. Before instructing drivers that certain
channels are no longer active, we use a 5-second timeout to reduce
on/off chatter. Also, we default to all channels enabled, but power
down after two minutes if no streams play within that interval. For
now these durations are in code, but this will be included in the
product-specific configuration in a subsequent CL.

The following items are still todo:
* Extract policy-driven constants (like kOutputIdlePowerDownTimeout
  and kInitialPowerDownDelay, and perhaps mappings from RenderUsages
  to frequency ranges) out of audio_core code, into a policy layer
  or product-specific static configuration file.
* If required by a device's turn_on_delay_, trigger
  OnMinLeadTimeChanged notifications for renderer clients.

Original-Fixed: fxb/82302
Original-Bug: fxb/75709
Original-Bug: fxb/75714
Original-Bug: fxb/82408
Original-Bug: fxb/82423

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/525788
Original-Revision: b9f7fc25392fa789adf52d879034a3d88749b372
GitOrigin-RevId: b351bc86388a2962d86cc1205dc34c7529e8b472
Change-Id: Ibe1c281e0dca863ea90e48d6d537ae1a7ca21197