[roll] Roll fuchsia [build] add support for "bootstrap" flash target
Bootstrapping a device is the process of taking it from an unknown state
to the point where it can flash Fuchsia using the normal flow.

This doesn't need to be done often - generally only when provisioning a
board in the first place or recovering a bricked device - but it will be
useful to have a standard flash target to support this flow.

An example of a file that may be included here is the GPT, which isn't
expected to change during the normal device lifecycle but does need to
be initially flashed once.

Original-Bug: 64549
Original-Bug: b/188193322
Test: ffx target flash <flash.json> bootstrap
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/567447
Original-Revision: 7ac01d47d8e27ab16964f27d88037edc1bedb08d
GitOrigin-RevId: cd5cc78c77ccc5af112a8ce1e4d26b1e7aae4e1f
Change-Id: I9a9e526d6c9b4e2c5063fc41e147d6570169f807