[roll] Roll fuchsia [benchmarks] Introduce runTestComponent() helper function
Introduce the runTestComponent() helper function and change a test to
use it.

This adds steps to generate summarised versions of fuchsiaperf files
and remove temporary files.  The former might change the results from
this test, because it means we'll now be properly discarding the times
from the initial warmup runs of the test cases (an issue I missed in
https://fuchsia-review.googlesource.com/c/fuchsia/+/559687 and which
is described in src/tests/end_to_end/perf/test/summarize.dart).

This function is similar to runTestCommand() except that it invokes a
component rather than an arbitrary shell command.

Original-Bug: 82465
Test: fx set terminal.x64
  + fx test --e2e network_device_microbenchmarks_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/568576
Original-Revision: 30a428019f49d6701c9baf1291f3d62a736cc619
GitOrigin-RevId: 4fe9c4419817ed9a7592a387f34bea5cb69ad14a
Change-Id: I6b6d64dae18d3ad72f2beafca557439d1bbf6fe4