[roll] Roll fuchsia [power] Replace CobaltMetrics with PlatformMetrics
Replaces the existing CobaltMetrics singleton with a more flexible and
useful node called PlatformMetrics.

The primary motivation for this transition is to support capturing the
same historical CPU temperature metrics on Nelson that we are on Astro.
There wasn't previously a clean way to do that considering the node that
publishes this metric on Astro (ThermalPolicy) is not used on Nelson.

This change restructures the CobaltMetrics singleton as a node and in
doing so, configures it to take a more hands-on approach to gather the
required metrics data. Instead of having CPU temperature pushed into it,
the PlatformMetrics node will poll the CPU temperature handler at a
configured interval (1s). This approach is easier to reason about and
eliminates the trap that we'd fallen into previously where we enable the
ThermalPolicy node for the sole reason of capturing these metrics, while
adding hacks to disable the actual ThermalPolicy behavior.

Test: fx power-manager-tests
Original-Bug: b/178334881
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/563144
Original-Revision: caeb81da175b35003118758652a0302bf5b95184
GitOrigin-RevId: f5ed70287fd38821afa42a339bb49ea1291cfa03
Change-Id: I1f7f269d26e6cb2d7c4792ebcc696411545175fd