[roll] Roll fuchsia [assembly] Use assemble_system in zedboot in parallel
We are using the assemble_system template for constructing zedboot in
parallel, and ensuring the resulting ZBI is the same as the current
assembled ZBI.

We are not checking the VBMeta's equivalence, because I got excited and
already yoinked the code that supports injecting a salt
(fxrev.dev/564324). The VBMeta is nothing special for Zedboot, so I
don't think we should be concerned with not checking for its
equivalence.

Original-Bug: 82474
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564848
Original-Revision: b0a3cadd5b30d0c2bad61fb2a97f5e363f8f4985
GitOrigin-RevId: 38fe1f5ef07053868883a467a740f9c5f304ab99
Change-Id: I93b1b3d8c4f5d2f1f7b7afdeeaa23310cfa9a7b5