[roll] Roll fuchsia [driver_manager] Move eager fallback to driver.cc
There are boot parameters that will mark a driver a fallback
driver as a normal driver. Now that a driver's fallback status
is set in driver.cc, this is also where the eager check
should be handled.

This means that eager/fallback checking happens in the same
way for drivers found through the DriverIndex or through
DriverManager.

TEST: fx test ddk-fallback-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/568181
Original-Revision: afa002f115c4f6a6eaedcf725882d0eee148bfbd
GitOrigin-RevId: 011917f966d4850afbbea4a2239a0d05be6eddce
Change-Id: I46d407c3eab0a4c2cb240c49ee2f773f72e3a893