[roll] Roll fuchsia [gn-sdk] Enabling auto indexing of package symbols
This adds a global flag `fuchsia_auto_index_symbols`.
When true, the ids.txt file generated from a `fuchsia_package`
is added to the symbolizer index using the `symbol-index` SDK tool.
The Fuchsia symbol index used to symbolize logs and to find source
when using zxdb.

The default value is false, in order to be backwards compatibility
with existing builds, and only makes sense to set to true when building
as a developer that will need to symbolize logs or use the debugger with
locally build code.

To enable it for a build add `fuchsia_auto_index_symbols=true` to
the gn gen args.

Bugs: b/195019412
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/567504
Original-Revision: 8848445523d45dd57fadaa517d63179228415484
GitOrigin-RevId: 83c3208fe2c10a367d2041e022903a9a584553e8
Change-Id: I00431ecc27e54206f5f9df5db845d096f6ac22ce