[roll] Roll fuchsia [schema id] update schema id of cmx_schema.json
In Cl 530463 I'd made a typo when changing the id for
src/sys/lib/cm_json/cmx_schema.json by omitting the src
directory, which is part of the path to the file in git.
This CL adds that "src/" part of the path.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/568987
Original-Revision: fb9155da2d6eb0f0772b564421992f95252c3c5d
GitOrigin-RevId: 5ee40858430f847136845ec6fab08850bc65b558
Change-Id: Id3d572da8e2b32f9df97e3b4f7869d6599aec86f