[roll] Roll fuchsia [lsc] Migrate BindChild to OpenExposedDir
This CL is part of a migration of usages of
fuchsia.sys2.Realm/BindChild to fuchsia.sys2.Realm/OpenExposedDir.
The former has been deprecated.

For this CL in particular, the log-stats integration test
has been changed to explicitly bind to the
`fuchsia.component.Binder` so as to start the `log-stats`
component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/563930
Original-Revision: f20d49ed157c44e1eebf4b53fd213f2785e143d7
GitOrigin-RevId: 9a755f97cbb957ab3193066fc57df1a66228e214
Change-Id: Ibfbbae9ff9ed33bbbdabbbf568ca44986878b787