[roll] Roll fuchsia [audio_core] Idle-power on output hot-swap
When an output device is added or removed when streams are actively
playing, IdlePolicy needs to account for any change in routing. Note
that in these cases, the number of active renderer streams does not
change, but an audio output device may no longer be the target of
active streams. This accounting is needed so that IdlePolicy correctly
manages the enabled/disabled states of audio output devices, not just
as AudioRenderer streams come and go, but also as other output devices
come and go.

AudioCore extracts the subset of RouteGraph that is used by
AudioDeviceManager, into the new interface DeviceRouter. IdlePolicy
implements the interface (and is injected into AudioDeviceManager
instead of RouteGraph), using it to take snapshots of the route graph
before and after devices are added/removed. Between taking these
snapshots, it passes the actual AddDeviceToRoutes and
RemoveDeviceFromRoutes calls through to the actual RouteGraph.

Also, debug-logging for the idle-power mechanism is reduced and
cleaned up, now that support for this feature has stabilized across
AudioCore and the audio drivers.

Two top-level build-time flags are added to enable/disable the entire
idle-power feature, so device power consumption can be easily
compared, before and after this feature. These can be removed in a
subsequent CL.

Original-Bug: fxb/75709
Original-Bug: fxb/75714
Original-Bug: fxb/82302
Original-Bug: fxb/82408
Original-Bug: fxb/82423
Original-Bug: fxb/82798

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/569502
Original-Revision: c9fe9b15c06c70f6bacfd636ac189cbcd95b80a4
GitOrigin-RevId: cd50c785a826e153b31379f883d7c20c64f807f7
Change-Id: I6861c4cb561304b199a7949a20b20e95e655cf38