[roll] Roll fuchsia [inspect][rust] Add counter for failed allocations
When we run out of memory and cannot add a new page with new blocks we
just return an error. In order to export how often this happens, we're
adding a counter that reports how many failed allocations we've had
since the heap was created.

Original-Bug: 75729
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/569421
Original-Revision: aa183fb3da715a4d489571983792ea8dd2355aae
GitOrigin-RevId: cec0bdb980611a35ec8f2b28cd7a7b81a699459b
Change-Id: I689274b9a346741253f5ca2fe3588111fb40398c