[roll] Roll fuchsia [pci] Handle MSI-X configurations sharing BARs
MSI-X capabilities can specify that their supporting tables may exist in
up to two device BARs, and that they may overlap with a BAR that a
device uses for data. Previously, we simply denied access to the BAR.
But now as long as we can ensure that the vector table and pending bit
array do not share a page with device data we can provide a mapping of
the space with that information excluded.

Additionally, a bug was found where inited_ was not being set.  This has
been fixed. Tests have been added both for Init as well as the new MSI-X
support.

Original-Bug: 32978, 80426
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/569223
Original-Revision: baff454743d3d48641ab9cf371f9c052060119d0
GitOrigin-RevId: 18015896824e52d662e6a78c3c877613c5a95075
Change-Id: I72d798540b7279d289c31d810ac2d64415beb503