[roll] Roll third_party/go [fidl][go] Use <= 4 byte unknown data to indicate inlining
Envelope inlining is a feature that allows up to 4 bytes of data to be
store inline in an envelope. Previously, the FIDL go bindings
identified inline unknown envelopes through a boolean on the unknown
value. This was unnecessary because it is possible to determine if an
envelope is inlined based on the data length. Additionally, other
FIDL bindings use the data length rather than an inline boolean so
this aligns better with them.

This also fixes a bug in encode where handles wouldn't be encoded in
inlining mode. This will be tested through GIDL tests added separately.

Original-Bug: 80348

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/third_party/go/+/569841
Original-Revision: 08436b6f74a8212cd4480cb5a1add7b71b84b622
GitOrigin-RevId: f8828923c348e816af2f82766e406527281f3df2
Change-Id: If32320ce23ec92e85657824c4ed6327621138129