[roll] Roll fuchsia [brcmfmac] Remove unsupported countries from list.
Original-Bug: 82462

As per the HW team and vendor, the following countries
are not supported yet for Sherlock:

KR, TW, IN, SG, MX, NZ

To be in sync with the clm.blob these countries need
to be removed from the list of supported countries.

Test: Existing tests cover the case of setting an
unsupported country code.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/569967
Original-Revision: 0ab84325334e825e9cd2ecca64b83a48dc9710bc
GitOrigin-RevId: a552a72d52c2bafe62ac1b708b2f9ba4ac6d6656
Change-Id: I71c8fbab5fa90441e252d1b09a96fd12e16bf15b