[roll] Roll fuchsia [ffx][repository] Work around `fx serve-remote` port collision
When using `fx serve-remote`, when both a local laptop and the remote
workstation may be configured to serve repositories. When this happens,
both repository servers may try use the same port. This currently will
report a warning, but otherwise not give any indication that the tunnel
failed to be created.

To work around this, this patch:

* Changes `ffx repository` to allow the port to be disabled.
* Changes `fx serve-remote` to detect the port collision, and error out.
* Changes `fx add-update-source` to report an error if the server is
  disabled.

Original-Bug: 82788

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/569743
Original-Revision: dc6528805b0188356dfb2446a094d7a4bf6f65a6
GitOrigin-RevId: 6926bf7671fd918579f1c1fe4f64d43dff6d76d1
Change-Id: I3ffe6787f8b4d79e47469074b9da338ec6ec98b5