[roll] Roll fuchsia [fidl][c] Change decoded representation for v2 unknown envelopes
The v2 wire format introduces an inline envelope representation in
addition to the existing out-of-line representation. As with the
v1 wire format, the decoded form of out of line envelopes are pointers.
Supposing that the decoded form of the inline envelopes were the same
as the wire format, the inline flags bit would overlap with the pointer
data making it impossible to discern whether the envelope is inline
or out of line. By shifting fields around in the decoded representation,
it is possible to have the inline flags bit overlap with the lowest bit
of the pointer, which is always zero due to alignment.

In order to implement HLCPP v2 decoding support, it is important to
have inline and out of line envelopes differentiated.

Original-Bug: 81995

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/569302
Original-Revision: 1df8b14ef5ff429b851ec66809abfab5d65df89c
GitOrigin-RevId: 3906d9dad7e852ab3d0b67b4822415bbb4c79254
Change-Id: I718100adc8f24c2087d4b6363853e68d11a3fa3b