[roll] Roll infra/recipes [luci_config] Checkout using git instead of jiri
In anticipation of moving away from Jiri for managing infra config
repos, convert this recipe to checkout using plain old `git fetch`
instead of `jiri`.

Conveniently, the `remote` property for the two existing builders that
use this recipe already corresponds to the repository containing the
configs, so we can just keep using that same property to refer to the
git remote, with no need for a soft transition.

The only downside is that the "lucicfg-infra-config" builder that runs
against sso://turquoise-internal/infra/config can no longer use the
version of lucicfg that's pinned in the prebuilt repo, because it no
longer checks out the prebuilt repo. So instead, just install lucicfg
directly from CIPD at a version pinned in the recipe.

This is probably fine because the configs in that repository don't
change very often, so it's rare than we'll need to atomically update
lucicfg and the configs, and not very disruptive to do that as a
breaking change. Eventually that repository will probably have the same
setup as integration where lucicfg is pinned in a cipd.ensure file
inside the repo, at which point we can remove support for installing
without a preexisting cipd.ensure file.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/570003
Original-Revision: 4f17e412942323e9a2df32278b85d50fd082255d
GitOrigin-RevId: 1cf104a1a7598c00f4a042c0dad2bad237a7174b
Change-Id: Ied7b773f577e754d6bc0c9189bfbc1a3b1f0ff0e