[roll] Roll fuchsia [recovery] Use rive instead of shed
In preparation for removing shed.

If an out-of-tree client has specified a shed file, this
implementation will work but will not draw a logo.

While there, fixed the fkey emulation of media keys that
was broken by the implementation of key repeat events
and improved the centering of the countdown numbers
a little bit.

Testing: manual
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/567382
Original-Revision: dd376accbffef3ad6036adc661ebbe496e311b23
GitOrigin-RevId: 4193297909f79adb59a91a8b14f8702674ac5bff
Change-Id: I7395230ecdacbeeb0e735a1c2da250945395b6e4