[roll] Roll fuchsia [realm_builder] improve realm destruction
Adds a drop implementation for `RealmInstance` that will spawn a new
async task that holds ownership of the mocks runner while it waits for
realm destruction to finish, such that mocks are stopped in an orderly
fashion instead of all at once.

Also adds a `destroy` function to `RealmInstance`, which can be called
to block until a realm is successfully torn down. Much like the new drop
implementation, this function takes care to keep the mocks runner task
alive long enough for the mocks to stop in an orderly fashion.

Original-Fixed: 78987
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/563576
Original-Revision: d78f6c0d8c63c1754690ffc4ade958d098cf1492
GitOrigin-RevId: 6432d08f7e982a9531b64e014a31f47bb0f27c03
Change-Id: Ie885b5700b9f6437ee4e6ce54fe6eef864c1cfcb