[roll] Roll fuchsia [hidctl] Ensure worker thread has finished before finishing destructor.
When HidDevice is being destroyed, we don't currently wait for the
worker thread to fully finish before completing the destructor. This
leads to races where the still-running worker thread attempts to access
state in the now-deleted HidDevice instance.

This doesn't appear to affect the driver proper (as the DDK will call
`DdkRelease`, which _does_ wait for the thread), but does trigger in
unit tests.

Update the class to ensure we join on the thread in the destructor.

Original-Fixed: 80427
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/570321
Original-Revision: 33a90886ec6a8e3b1e5dbeca53005480d6cee826
GitOrigin-RevId: d4558a808b298379417f59defa479546b422dfaa
Change-Id: I9353091d1368112ae576dc9295f7aa9f4606aa63