[roll] Roll fuchsia [archivist] Remove unused call to Tree#GetContent
We used to do this so we held a handle to the root VMO of a component's
inspect data so that later we could use it for recording post-mortem
inspect when a component is stopped. However, at the moment we are not
doing using it for anything and it doesn't seem like we will anytime
soon. So for now, let's remove this unnecessary FIDL call that we do
when reading data from most components.

Tested: existing tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/569925
Original-Revision: 5382ddb6f6a2a9aa86cc062b1026f458be8ff2d5
GitOrigin-RevId: e36dbdd3ee1f3567ae25c0d6fa56d9cd28ea473d
Change-Id: I924b4c3f72fcac220617d10c1b61b9cd3cc16dea