[roll] Roll fuchsia [hlcpp] Use correct type shape version in for request allocation
This may have broken out of line objects in compatibility tests.
No test is included in this CL because it is a small fix that
I want to get in quickly, but I am planning to make better
integration testing of the new wire format a focus in the near
term.

Original-Bug: 81995

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/570402
Original-Revision: b7ab6a9686f430ffcb36e86762002faff37af3cf
GitOrigin-RevId: 5151714ca858d9efe93375b3f9468024dc8f2682
Change-Id: Ia46f3aceb622cc70cea555401daa758fb0919a1e