[roll] Roll fuchsia [ffx] remove outer mut from generated macro code
The ffx_plugin macro generates a wrapper function which then calls into
the function which is given. That wrapper function does not have the
same mutability requirements as the actual implementation of the plugin.
Therefore you can get into situations where the compiler warns you about
an unused mut in the generated code because you need a mut in your
implementation. This change removes any mut from the command argument to
the wrapper function while preserving it in the implementation.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/570304
Original-Revision: c58855d51189f9fcbc6271d776d49b486cecc5aa
GitOrigin-RevId: 4cd3f77b213935757700f7f81a2fe467f088bd8f
Change-Id: I6c581fcceb138cec2ba7b864c840f3e614934e49