[roll] Roll fuchsia [zerocopy] Add transmute! macro
Add the `transmute!` macro, which is a memory-safe equivalent of the
`core::mem::transmute` function. The `transmute!` macro requires that
its argument is `AsBytes`, and produces a value of a type which must be
`FromBytes`. The two types must be of the same size.

The reason that this needs to be a macro and cannot be a function is
that there's currently no way to express in the type system the
requirement that two generic types must have the same size. The macro is
able to make use of the fact that `core::mem::transmute` (which
`transmute!` uses internally) is special-cased in the compiler, and can
thus use compiler magic to ensure that the two types have the same size.
As a result, `transmute!` can only be used in a context in which both
types are concrete types rather than generic type variables.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/569463
Original-Revision: 14b5ba00c9c19671b7b4e55f03f0c000d145a7ea
GitOrigin-RevId: ec618137d3b01116ce2c4baca086fad3b7193441
Change-Id: I4b7039f4e4a646ed3abb5ede411c069807cff8d2