[roll] Roll fuchsia [hlcpp] Fixes needed to get compatibility test working
These are fixes to two bugs that when fixed result in the FIDL
compatibility tests passing for all compatibility test servers
except for go and dart, which fail some of the tests.

These bugs are oversights that are difficult to cover in a
useful way by non-integration test tests.

Original-Bug: 81995

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/570861
Original-Revision: 5332db5dc3d81ed8ea8e75ec849232ae073a3843
GitOrigin-RevId: 97fe0dcec5f24b98fa8fb7e63eebae1a5ed52365
Change-Id: Icee567a9e7072e57c5e9b7364188d12b2266d213