[roll] Roll fuchsia [ffx] Implement `ffx component create`
`ffx component create` is the advanced version of `ffx component run`.
Developers specify the precise moniker of the component they wish
to create along with the component URL.

The `create` command will connect to the Realm protocol in the
incoming namespace of the parent and use the protocol to create
a component instance.

Common code for `create` and `run` has been migrated to the common
library.

The laboratory shard's `expose` decl for the Realm protocol has been
changed to a `use` decl. This allows it to also be used with
`ffx component create`.

Test: fx test ffx_component_create
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/562621
Original-Revision: d33871e93e48aa7230d5b2ae67fd67a729be547f
GitOrigin-RevId: c80ecb831ba6fcc842ace24292d510171c20618d
Change-Id: I871e8d0f43ed882b72617a9b6092fac6d917854b