[roll] Roll fuchsia [assembly] Use sdk manifest to find tools
The fvm, minfs, blobfs, and zbi tools are now located using the sdk
manifest. When building in-tree, the manifest points to host tools
available in `out/default/host_x64/<tool>`, but when run from the SDK,
the manifest points to the tools in the SDK.

This commit also refactors how we handle tools in tests. We now have a
helper function which can generate a fake tool at a specified path. The
tools are faked, and the paths to the tools are passed down to their
respective builders for usage.

Original-Bug: 76378, 79072
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/570097
Original-Revision: 85ee14edc8606415d4503efd9dd5679ac5197eb0
GitOrigin-RevId: e9e364cda5af246bf3de7e4a3942c10787a46414
Change-Id: I5407c05dea3b76729a21e320cb571bdabab8f013