[roll] Roll experiences [ermine] Enlarged the touch target size of the close button on a button chip
Since the touch target size of the close button on an AppChip was set too tight, it was hard to press it, especially by tapping it with a finger on a touch screen. Now the button has the default touch target size so that it can be easily hit.

Screenshot: https://screenshot.googleplex.com/NsSNWxYmrWJjyh2

Original-Fixed: 82529

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/570723
Original-Revision: e07e2679d8e5dc6d91a6d9d89f6bc38030518e7e
GitOrigin-RevId: 298caeeb0f90cf904b62aca1e268a54887ffcb82
Change-Id: I8376df2033edaf769b6c468cdd970bbcd481d3d8