[roll] Roll fuchsia [wlancfg] Runs regulatory manager until failure.
This change allows the regulatory manager to continue
running in the cases where setting the regulatory region
fails and when a country code of an invalid length is
encountered.  The loop will still terminate in the event
that the regulatory manager can no longer receive updates
from the RegulatoryRegionWatcher service.  Such a failure
will be propagated to the wlancfg main.

Original-Bug: 82663
Test: fx test wlancfg-tests
Multiply: wlancfg-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/568567
Original-Revision: 9000adab30aad78088828de10b855d6af1492621
GitOrigin-RevId: fd01720fcdcb31824ac3bf8b29e7166229c8a4b9
Change-Id: I6926341226d71448d4a40db11bed4d582b04e5d6