[roll] Roll fuchsia [swd][security] Add policy labels import to core.gni
Adds an import to make policy_labels available in core.gni for leaf
product definitions. Does not define a policy_label in core.gni yet.

This is essentially a no-op since in //build/security/policies.gni, the
policy_label arg is an empty dictionary. This is part of step 3 in the
SWD policy migration.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/571311
Original-Revision: f10199a7947859f31f6e4a16759fdc2fea159993
GitOrigin-RevId: 46e0fa034f58e7e923ea165f8b42498188ca1c4c
Change-Id: I1cd49ec7ab19e6d056d8177becb175f318cbd85b