[roll] Roll fuchsia [ffx] remove unused mut from target flash
It turns out this mut is just really unused and is not a spurious issue
related to the plugin macro. If the implementation was not split out
into a separate function then this would be an issue. With the
implementation being a separate function with its own mut declaration on
the command parameter, you end up not needing the one at the top level
as you are simply moving that value into the called implementation
function.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/570305
Original-Revision: 4e4d115e0c79ba394529a53d35d23faac3c8980c
GitOrigin-RevId: 3ef4f78a3dfcccf486978bd5c50c18fbd4e19907
Change-Id: I7ec59f986f22dd50a80fd56fbf8513c6403b980f