[roll] Roll fuchsia [vboot-fwparam] Add cros-nvtool.
This is a subset of the "crossystem" tool on ChromeOS. The crossystem
tool gathers data from several sources, including nvram, and allows
interaction with them. cros-nvtool is similar, but it only interacts
with nvram.

This is a raw binary that gets installed to bootfs so that we can use it
from zedboot. This CL also adds all the necessary plumbing to give the
various shells (serial and ssh) access to the `fuchsia.vboot.fwparam.FirmwareParam`
protocol.

Original-Bug: 82432
Test: cros_nvtool get && cros_nvtool set dev_boot_usb true; reboot and
      check the settings stuck.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/569203
Original-Revision: 8b3bcfed7aebdb699d1748bf4b54e4a2c35f6a98
GitOrigin-RevId: ae21a55bc82a02d483f1d51b224666d5eabff87b
Change-Id: I4961cb0b54d60a9071b67846744715073e2a509e