[roll] Roll fuchsia [f2fs] Land an initial version
This is an initial version of the f2fs landed in the main Fuchsia
as a copy of //third_party/f2fs. Since it does not have any changes
made in source and build files, it doesn't hook up with Fuchsia yet.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/572102
Original-Revision: 5cfd015a19bad8c485a971a51472a3c04fd24e03
GitOrigin-RevId: e62dcaf3cf4358c5ad14329b24ad9e730cfeb1c9
Change-Id: Ia6cdccc1b1a83793543a9e6fb53e4a9c1a2fb7f7