[roll] Roll fuchsia [rust][ffi] Create a basic example of cxx usage on Fuchsia
This is a modified version of the demo from
https://github.com/dtolnay/cxx, adapted to demonstrate usage of cxx in
the Fuchsia tree and usage of the generated from both C++ and Rust.

This also serves as a basic test of the rust_cxx_ffi_library template
and cxx integration in general.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/562423
Original-Revision: 6b9e9399dda33e9b5a6b9fe06388af3a4ffcc6b6
GitOrigin-RevId: 2b869e21519b91abbadb599d1ad5803dd017f0c9
Change-Id: I226d84e0f3cf8f1e903d5425918ed83e0eaf9175