[roll] Roll fuchsia [tpm] Move command handling to a thread.
This CL moves command handling to a thread in preparation for taking
requests from other applications. Commands are submitted to a queue which
contains the command to send and a callback to call on its completion.
The callback is called with a zx_status_t (in case command transmission
went wrong) and a TpmResponseHeader* (which is the result from the TPM).

The TPM now no longer performs blocking operations during suspend/init
and command responses are passed back to the caller of the command.

Original-Bug: 81433, 82504, 82505
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/572722
Original-Revision: 921137ee8fd8d91404078140aa0e942a551c1689
GitOrigin-RevId: 0ff9668a701b83a2afb750fb38e4b9c23cd4ae39
Change-Id: I9bcd381425c7e5220554072e8a4824a21f60675d