[roll] Roll fuchsia [utest] Attempt to remove flake in PortTest::CloseQueueRace.
We currently see a hang in the test in CI/CQ roughly once per
month. The race appears to be:

  * Thread A enqueues a message on a port.

  * Thread A decides there are too many messages on the port, and
    attempts to dequeue a message by calling zx_port_wait.

  * Inside the kernel, the handle to the port is converted to
    a reference to the port object, but the queue of messages is not yet
    inspected.

  * Thread B deletes the handle to the port. In the process, all
    messages on the port are deleted.

  * Inside the kernel, thread A determines that no messages are
    on the port queue, so it waits forever for one, causing a hang.

We resolve this by using an infinite-past deadline for the zx_port_wait
operation, and trying to read a message again if a ZX_ERR_TIMED_OUT
error status is returned.

Thanks to maniscalco@ for spotting the cause of the hang.

Tested with 100,000 iterations, 15 of which triggered the
ZX_ERR_TIMED_OUT race.

Original-Bug: 83334
Original-Fixed: 82914
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/571623
Original-Revision: 3297c77b01ebb038426e4a15af3809d51f0558d0
GitOrigin-RevId: 34dcd9d73f2515a702ef9378f380ea1bd15224c0
Change-Id: I0c8014a82c24ff3d46d7702a638351791108202b