[roll] Roll fuchsia [kernel][preempt] Clear local preempt pending in int_handler_finish
During the call to int_handler_finish at the end of exception
processing, if preemption is allowed after dropping the preempt
disable count added at the start of the interrupt handler, and local
preemption is enabled, clear the local preempt pending state
immediately, instead of waiting for it to be cleared later on in the
scheduler.

Doing this prevents a window of opportunity in which the system might
perform an extra preempt operation (at best), or suffer from infinite
recursion (at worst), if preemption is disabled and re-enabled again
after the call to int_handler_finish, but before the preemption
operation has been completed.  See fxb/83079 for details.

Original-Fixed: 83079
Tests: verified using repro steps described in the bug.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/572604
Original-Revision: e609ca44e2411f49d3094b908fded0202c790531
GitOrigin-RevId: 655ca9c48d18e6d4a2e06dfdfe532dfcaf970e0d
Change-Id: I4a97939c934a766c11e0bde872067cee3bc3b566