[roll] Roll fuchsia [identity] Remove redundant Identity API methods.
This is a first in a chain of CLs to adapt the existing identity API
surface to the simple authentication prototype that we're in the process
of building.

Step 1 is to remove the methods that are now out of scope for Fuchsia,
namely deep intregration with cloud based identity services (this can
be added by products if needed) and for a global notion of an
multi-device account (again products could create this if they needed).

Original-Bug: 81605
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/570825
Original-Revision: 5d56fa1e4fe6024e843b66a318daa5316b52f580
GitOrigin-RevId: 2d214ead9c64838d93fdaf63a15f3380da93598a
Change-Id: I99b2561bf7c195aaa3385ada03ca89b3d22a5aa0