[roll] Roll fuchsia [cts] Don't verify deps when releasing the SDK.
sdk_atom.gni (the basis for every SDK rule) will now save the filepath
for every file in the active SDK in a metadata field named "cts_artifacts".
We use that metadata field to define the CTS SDK, and infra uploads
the files in that list to CIPD.

When releasing the CTS SDK, we need to make sure the dependency tree
only includes things we want to vend to CIPD. So we have to skip this
verify step in all of the rules.

This is OK because the deps will be validated during regular developer
workflows (when the CTS isn't being released).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/573369
Original-Revision: 8cf247f2747afdd493bdcbe3027ef27248e1267b
GitOrigin-RevId: b2a901a87963ac3b7eebc6f8e176dfa717fe988a
Change-Id: I4b8a29d94290790d680fa9943c5c6556c3d3ad2d