[roll] Roll fuchsia [swd][security] Apply swd policy labels to core.gni
This is part 3 of the process to migrate to using policy labels for SWD
as described in RFC-0118.
Prior steps can be found at: fxrev.dev/561861 and fxrev.dev/566505.

Manually tested by running `fx build` with and without the change to
ensure that ninja found no work to do, which by my understanding means
that the cumulative set of targets has not changed.
Verified this with the following product:
- core.x64 (core.gni)

This product definitions were found to be the only one that had
references to the SWD targets that are to be migrated to be encapsulated
by the policy definition. In addition, a codesearch was done to find any
gni definitions that defined its own `base_package_labels` or
`product_system_image_deps` instead of merely adding or subtracting
packages from the inherited lists. None were found.

For reference, the migration process is as follows:

1. Define all SWD policies.
2. Add dependency on SWD policies in base_package_deps and
   system_image_deps, but do not apply a SWD policy to retain existing
   product behavior.
3. Apply SWD policies in product definitions alongside the existing
   SWD targets configuration. Assert that product configuration has not
   changed.
4. Remove manually defined SWD targets in product definitions, rely
   solely on defined SWD policies.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/572049
Original-Revision: f18e3d38851a54d29fe49e08d9b6716231c65d70
GitOrigin-RevId: 9fdb12da0733c9f6ca8c2c43c65dbe39b30ac692
Change-Id: I8527293f148a2da5c05abd31545ec87ae2795ac3