[roll] Roll fuchsia [scenic] Make the gfx::Scene object trigger on_view_created_
This is a workaround for the root session effectively being capable of
having two views (in the view tree): the scene and the view.
Because the scene object is the only guaranteed singular root of the
scene graph we want to report that as the "view" of the root session.

This CL make it so that the first scene or view object created by a
session is what is marked as the session's view when registering
protocols with CreateSessionT().

In theory this could prevent the root session's non-focus protocols
from working correctly, but all the current ones are useless for the
root session anyway.

Original-Bug: 82643

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/570848
Original-Revision: 51512b98924735d8a4c3d077257ac92464c5d7c4
GitOrigin-RevId: 54379c079592324acc761bd45ccd31d936158441
Change-Id: I9e03248ceeee3ad3a1c50853379aacd1ad6ce9e2