[roll] Roll fuchsia [lowpan-spinel-driver] Do not bring up net iface when commissioning
This change makes sure we don't bring up the network interface when
we enter the commissioning state. We don't actually have network
connectivity in this state, so we shouldn't imply that we have
network connectivity until we are actually provisioned.

Original-Bug: 82069
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564584
Original-Revision: aa05fe34cbe5002831cd4dc15819ad717abd61b4
GitOrigin-RevId: f4a59bdb5fb26e6ebf726a609b1a1ea8ad2f7890
Change-Id: Ibe8a6a3c07f2a84114b40f37f26de40a30701cb9