[roll] Roll fuchsia [storage][tests[: Run in isolation.
The previous situation would keep the component(isolated devmgr) as part
of a group of packages that contained the storage tests. This means,
that if the user added the group, the test dependencies couldbe
resolved.

Now this change, allows a user just adding package A(not the group), and
have this package contain all the components it needs to execute
successfully.

in fx args the situation would be:

Before: universe -> [my test, isolated-devmgr-component]
Now: universe -> [my test]

Test: fx test fs-managemente-tests (and CQ)
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/571445
Original-Revision: 32cb74f16e03de7abd6f9f9237300fc6615ce7da
GitOrigin-RevId: 57c69736ccde26742bb6e7429976c885ce2108e1
Change-Id: I3d5d0995e0f408319625474c8236818c9dd644c9