[roll] Roll third_party/go [fidl][go] Part 1/3: Remove Marshal() and Unmarshal()
There have been a number of bugs from not providing a MarshalerContext
and it isn't much more verbose to add one explicitly. Because of this,
it makes sense to remove Marshal() and Unmarshal().

Part 1 does the following:
- Expose NewCtx() so MarshalWithContext() can be called with a default
context.
- Change the signature of UnmarshalWithContext so we can migrate to that
from UnmarshalWithContext2 (for a simpler name)

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/third_party/go/+/572981
Original-Revision: 7ae640525080564d96d8759b945b1dbf7543717c
GitOrigin-RevId: e6f4e498b93d2b13bdb14f18ef1a56063a1879ea
Change-Id: I9e6d42d9196d4c8d7de5cea8e46fc6db948afbc4