[roll] Roll fuchsia [kernel][oom_killer]: Fix sort order.
Originally qsort was used which takes a comp function that returns negative zero or positive, depending on the relationship between a and b, on comp(a,b).

That is if (a-b < 0) implies a < b, a- b == 0 implies a == b and a - b > 0 implies a > b.

CPP stable sort takes a comp function that returns true if a < b, to be read a goes before b in the collection.

The implementation was kept but the meaning changed.

qsort comp function was intentionally written like that, which would produce the descending order expected in the loop following the sort.

On the other hand, stable_sort(while keeping the same comp function) produced the ascending order, which breaks the expectations of the loop following the sort.

The test caught a bug, but the way the code was writing relying on an implicit bool to int conversion of a relation operation caused the problem.

Also the problem might have gone unnoticed, since the existing behavior is to reboot on OOM.

The fix is simple, flip the relationship operator.

Original-Bug: 45616 #Done
Test: k ut job
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/569666
Original-Revision: 54202c25ac655524d223655d17bbbd1a592b55cc
GitOrigin-RevId: 3813fa70358683b600a4164008aef93d9f21ceaf
Change-Id: I0b8a0935e8b228053cbf7d83414e968fede52254