[roll] Roll fuchsia [cts] Run generate.py to produce the finalized CTS SDK
This CL runs generate.py on the CTS SDK, which pulls in the needed
dependencies and adds BUILD.gn files, so we can properly build the tests.

In the future we will want to transform the CTS SDK _after_ we retrieve
it from CIPD, but we can't do that in CQ yet.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/573562
Original-Revision: 7f3950113bfe8e77a42f658626d5d20c70d03def
GitOrigin-RevId: 48d187f37815d27b8886c6ef3efc2b9c9a9a4073
Change-Id: Iabc4141779e3ebec41fae77e68305a07dd1804ef