[roll] Roll fuchsia [lowpan-spinel-driver] Increase join timeout to two minutes.
In commit [87704169d0f][1], we increased the join timeout from 15
seconds to 30 seconds. However, testing has confirmed that 30 seconds
is not long enough for all devices. As a result, we are increasing this
timeout to the theoretical maximum: two minutes.

[1]: https://fxrev.dev/Ie6db4a2767232cbd3029338dc0624a9409231814

Original-Bug: 83525
Original-Bug: 82738
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/573950
Original-Revision: 806d54000fef94be520b54dc337a902335314fa3
GitOrigin-RevId: 790aa7e97b42ce9196abf7e5a9dae68318955235
Change-Id: Ib67b2088880885204a8a4b71e1068ea63f75c9f1