[roll] Roll fuchsia [hlcpp] Cleanup of hlcpp-focused variants of fidl_decode
Currently, it isn't obvious that the *_skip_unknown_handles
variants of fidl_decode are specialized for hlcpp and the
new naming makes this clearer.

In addition, now that it is clearly hlcpp-focused it is
possible to remove unused variants of the functions.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/572947
Original-Revision: 532d0c23c7331afb886e8814c28545c9e5d97be9
GitOrigin-RevId: bb4e96a435df9f19400cb3d71f164679bdc564a6
Change-Id: I69efb64797c3f12e4be48d47d1a3ed0a2e26ac29