[roll] Roll infra/recipes [fuchsia] Propagate failed subbuild output props
Previously we wouldn't propagate most of these properties from the
subbuild to the parent build if the subbuild failed. This was not
intentional, as `fx repro` and various other tools that are often used
even for failed builds rely on these properties being set.

Now we set the output properties on the parent build even if the
subbuild fails, making it possible for `fx repro` et al. to retrieve
them even for failed builds.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/573872
Original-Revision: deb00e86e92eeb177c6c0b54037bbe200118debe
GitOrigin-RevId: 02d74b53fce3ba194b1bd92d05eb48608bd33798
Change-Id: I82a2f8c6fead881e3945f7702de17c54541ac41e