[roll] Roll fuchsia [wlan][policy] Remove active scan decider
This change removes the active scan decider concept in favor of
a purely probabilistic active scan decision.

The hidden network probabilities are now updated after passive scans
so that the most up-to-date information can be used to make an active
scan decision. If an active scan occurs, the hidden network
probabilities may be updated with those results too.

This means that, in practice, there is no way for a caller to
deterministically request a passive-only scan while there are saved
networks, because the decision to active scan is based purely on hidden
network probabilities.

This also required some test refactoring and mocking to maniputlate the
hidden network probabilities to ensure scans ARE deterministic for
testing (i.e. setting the probabilities to 0 or 1).

Multiply: wlancfg-tests
Test: fx test wlancfg-tests
Original-Bug: 75668
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/557728
Original-Revision: f85e2e96aae93cb20b54ff00f3bbd3618a5a3000
GitOrigin-RevId: 0c505cbb42cdc3ecfb3840fb44894c72f5f5a045
Change-Id: I3c967910f7bea814c62b9d45153c0de772d19598