[roll] Roll fuchsia [fdio] Add inotify tests that run on Linux and Fuchsia
inotify is a Linux interface that Fuchsia supports through fdio. This
adds a number of tests for the fdio implementation, including several
that run on Linux as well as Fuchsia.  This also changes the error for
an empty path to ENOENT to match Linux.

Original-Bug: 60109
Original-Bug: 73151

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/574048
Original-Revision: 5ba70c1f67e8a39e9ef7e6504d947ee9ad81b054
GitOrigin-RevId: 048e9a4df6c639fc271fb8543a84eb1bcab84f84
Change-Id: I944e4dc53c86dba1115a6588b0490eaf468a2122