[roll] Roll fuchsia [kernel][crashlog] Refactor the platform crashlog API
Refactor the platform crashlog API just a bit in preparation for a
change which will allow crashlogs to be rendered directly into
mappable NVRAM regions, eliminating the need for statically (and
sized) rendering buffers.

There are two main changes to take note of here.  First, the largely
C-style API, along with its individual function pointers, have been
replaced with a C++ vtable indirection approach.  The system starts
with a trivial crashlog implementation which does nothing, and this
trivial implementation may be replaced later on once the crashlog
capabilities of the system are understood.

On ARM systems, this is always a mappable RAM based implementation,
while on x64 this might be either a mappable RAM based implementation,
or an EFI capsule based implementation, depending on what is
discovered in the ZBI.

The second change has to do with how the render buffer is accessed.
Previously, the buffer was a statically allocated 4k buffer located in
platform/power.cc, and used in platform_halt.  Now, the buffer is
fetched from the crashlog implementation by platform_halt instead.
Once the crashlog has been rendered into this buffer, platform_halt
"finalizes" the crashlog, allowing the implementation to take the
steps needed to sequester the final version of the crashlog to survive
across a reboot.

Tests: manual testing on NUC and QEMU ARM64.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/568403
Original-Revision: 9a90d584684045794b49c45e9b61dd6d0b6a872e
GitOrigin-RevId: d83f11620d9f0591da174c68284e111f28b26350
Change-Id: Ic7fc32e6412c252583c138a800c6290017291d46