[roll] Roll fuchsia [starnix] Fix symlink accounting
Previously, we were resetting the number of symlinks we would follow at
each step of the resolution process. According to LstatELOOPPath, we're
supposed to apply the symlink budget to the entire resolution process.

This separates the notion of the SymlinkMode from the LookupContext. The
SymlinkMode now just records whether we intend to follow the symlink at
the end of the resolution process. The LookupContext records the number
of remaining symlink follows. As path resolution proceeds, the number of
remaining symlink follows in the LookupContext is mutated, which counts
all the symlinks follows from the whole lookup call tree.

Test: StatTest.LstatELOOPPath
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/574056
Original-Revision: 31063d2f81f3712c9085bc9fa885f36e80632b44
GitOrigin-RevId: 3d5a66a2aa4617bca07bafc9d48407331fff5851
Change-Id: I652fac34c4a904da5728c5b6f6cb9a935f6a2d81