[roll] Roll fuchsia [kernel][vm] Retain non-terminal accessed bits
This changes accessed harvesting to not remove the non-terminal
accessed bits if we are not planning to free them. The purpose of this
is to avoid destroying information so that a later FreeUnaccessed
action behaves the same, even if there were 1 (or more) Retain
actions in the middle.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/572109
Original-Revision: 34de61bf5d8e2e211dfb42804936fbdde5bfccb5
GitOrigin-RevId: 4e7400638d4410e214fe10b10996c5e79d8480cf
Change-Id: I4bfaad756c04e70eb0cdc9d4d9d6a89a9138ba76