[roll] Roll fuchsia [carnelian] Support flatland
Add support for the Flatland compositing system.

Added a new view strategy for Flatland, much of which is a copy of the
strategy for scenic. Since the scenic strategy is going away with
scenic, it might not be worth the effort to share the common
implementation. Reviewers are encouraged to suggest otherwise.

This implementation works, although the lack of stability in Flatland
makes it hard to be sure this support is complete. Ten second traces of
rendering do look very solid, though, with no missed frames. Still, it
seems worthwhile to commit this CL as a basis for further efforts on
Flatland support.

Original-Bug: 83510

Testing: manual

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/569921
Original-Revision: 9d51ca10ab2d82f954632c384573d3ae5b875fc4
GitOrigin-RevId: 27c3d7de0850127a09c18bd4db7805e5c759e5c9
Change-Id: I0fcd631def12d5e936414b0a69e07c1bb146040f