[roll] Roll fuchsia [ffx][config] Remove RealPaths
For Sdks built from a sdk directory, the prefix was being appended
twice, causing the path to be incorrect. We do not need the Prefix type
for RealPaths, and can replace it with a no-op. And to clean it up even
further we can remove RealPaths altogether and use an optional map.

Original-Bug: 83609
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/574204
Original-Revision: 01666b2822c56df7a72d0a16045961d444fcc6e5
GitOrigin-RevId: 31f6caa1af0cd3a3cc534c6c01de06c7e7ab7e05
Change-Id: I388b49de524a899182d6e50fce4394dfc84af22f