[roll] Roll fuchsia [kernel][hypervisor] Ensure vmclear is run on correct CPU.
AutoPin is a class intended to pin the current thread to a particular
CPU. It is currently only used during Vcpu deletion to ensure that
vmclear is executed on the CPU that last ran the Vcpu.

However, due to a bug, AutoPin doesn't actually migrate the currently
running thread to another CPU. Instead, vmclear is always run on the
current CPU instead of the CPU the VCPU last ran on.

This CL replaces the single use of the AutoPin class with
a mp_sync_exec() call that executes the vmclear on the correct CPU.
Additionally, we add a new test case exercising this path.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/573204
Original-Revision: 9aeca66429787fa90af3fbd9725a18bdeddcdf17
GitOrigin-RevId: 1c21ff09795326796bde100c8b485afc984e3382
Change-Id: If572928662b408a9f8cb82b744b84967e17a69fb