[roll] Roll fuchsia [cm_rust] Make source_path optional in CapabilityDecl
In a follow-up, we'll use CapabilityDecl to register built-in
capabilities as well as component and namespace capabilities.
Since built-in capabilities don't have a meaningful source
path, this field is now optional. The validation for
component and namespace capability declarations still enforces
a non-None value.

Original-Bug: 83626
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/574406
Original-Revision: 32f16ed38562187af1b95408fb74a146dff04924
GitOrigin-RevId: a5ea7a4f36886ec372f13e4429dec181f1a4d909
Change-Id: Ib9eb968c25eddd6a73b2e25ea74a970c62beeba6