[roll] Roll fuchsia [test][cprng] Add more cmdline entropy tests.
This CL adds two more tests that exercise the cmdline entropy behavior.

Basically, one test invokes the emulator twice, using different cmdline
entropy, and runs a binary that prints random numbers. If the cmdline
entropy is being processed correctly, the set of random numbers must be
disjoint.

The other test is similar, except that it uses the same cmdline entropy
on both invocations. The expected result is that the random numbers
overlap at least once. This is used to validate that the first test is
indeed using the cmdline entropy as the only source of entropy.

In order to use only one source of entropy, the CL adds two boot
options: `kernel.cprng-disable.jitterentropy` and
`kernel.cprng-disable.hw_rng`, to disable external sources of entropy.
Note that the global cprng still honors the cprng-seed-required and
cprng-reseed-required flags, meaning that if any of those are set, and
the source of entropy is disabled, the kernel will panic.

These tests assume that nobody else uses the cprng (we only have one
userspace process, and the kernel doesn't use it), and that the uses
that happen before the first userspace process are all the same.

Test: fx test kernel_cprng_tests
Multiply: kernel_cprng_tests : 200
Original-Bug: 82809
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/573144
Original-Revision: 0fcf32ced8563e90eed36310d2cc67d4cf3328a8
GitOrigin-RevId: 9fee8ad00d89cc582d6eb499fe72452cebf58301
Change-Id: Ib2370e9ae3116573bba8992ef11f7b878d235cca