[roll] Roll experiences [ermine] Remove mobx from AppState
MobX observables can be moved out of AppState, which is an abstract
class. It is then an implementation detail, handled in
app_state_impl.dart and hidden behind getters/setters.

This also has the benefit of removing all '.value' dereferences, which
were leaking the implementation detail that object is an observable.
Instead of exposing MobX Actions, we just provide simple methods that
are more typesafe because, unlike Actions, their parameters can be
passed directly instead of in an Array.

ViewState and SettingsState will be updated in subsequent cls.

Test: All updated unittests pass.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/574153
Original-Revision: e1dacbc5537c3b12e644f183500113b250ec91ae
GitOrigin-RevId: 2a5c0f2e0b521ac75ed6499d6f00083f6489ed3a
Change-Id: I04b3896edf75a241ebcb74f29e3270065d68aaaf