[roll] Roll fuchsia [wlan][iwlwifi] try to load the next version file if unavailable
This is a follow-up of c1df548d8ce4 where we missed the retry
mechanism.  In this CL, the code will try to load the next version
if the current version file is not available.

Original-Bug: 76121
Tested: On 9260, the max version is 44 but the available firmware
        version is 43.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/562782
Original-Revision: 27f70b96afdd0216e33fce4e05dc00a58cfca839
GitOrigin-RevId: 8ee381d97ba4aa56f206e199e79490e3e26d6b43
Change-Id: I3648d8e7f0036c240480557ace3269ec0136e276