[roll] Roll fuchsia [virtualization][arm64] Load Linux kernels without EFI support.
ARM64-based Linux kernels can be compiled with or without EFI support.
In both cases, they require a header as described in the Linux kernel
docs [1].

When a kernel is compiled with EFI support, it additionally requires
a "MZ" prefix at the beginning of the header, and some additional
reserved fields filled in, as described in [2].

The VMM doesn't require EFI support to be compiled into the kernel, and
can happily load kernels without it. Remove the MZ header checks.

[1]: https://www.kernel.org/doc/Documentation/arm64/booting.txt
[2]: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/Documentation/arm64/booting.txt?id=4370eec05a887b0cd4392cd5dc5b2713174745c0

Original-Bug: 82425
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/574812
Original-Revision: 3a93c1e564fb38adf4f5ac9ac219da85862f330c
GitOrigin-RevId: b6f6133ab091b7fa0c69bb3e3b8d17c149c0f949
Change-Id: Id1e7564c69811c1d7f99c4e6a336ae2db38260eb