[roll] Roll fuchsia [fidl][llcpp] Move generated bindings
This moves the generated bindings from:
  #include <library/name/llcpp/fidl.h>
  #include <library/name/llcpp/fidl_test_base.h>
to:
  #include <fidl/library.name/cpp/wire.h>
  #include <fidl/library.name/cpp/wire_test_base.h>
and the fidl.cc to wire.cc too.

The plan is for the unified C++ bindings to all reside in this same
location.

This CL includes a forwarding header for the old llcpp/fidl.h location
but and the llcpp/fidl_test_base.h. The use sites will be updated and
the forwarding headers removed in future CLs.

Original-Bug: 74067
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/573485
Original-Revision: e9003169ed1faa22b3a2802a5dec3e9e5089858d
GitOrigin-RevId: cf1cc4ec70f1340fdd271addc19d2ec8410b7792
Change-Id: I74c4bf5ff9ee46d2d0d81d71cbf5759959ddfcf0