[roll] Roll fuchsia [fidl] Add a comment to clarify light sensor data lux value to be 'clear'.
And delete old comment since it's wrong in current Nest light sensor chips.
The lux field from the light sensor is actually the clear value, it is not lux. Any use of the lux value has to be used as the clear value. It is incorrectly named.
Possibly deprecate and make new field with correct name in the future.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/572525
Original-Revision: f6dd2d0dd61905c7308ec6cc11ff26679af0f06a
GitOrigin-RevId: 3c29e0af8eba68e93fca1273cad296bd1159629b
Change-Id: I6396e48cf41661f1b80ed18832d3a96a9c601248