[roll] Roll fuchsia [kernel][test] Add test exercising singleton mutexes.
The test ensures that acquiring a singleton mutex is thread-safe the
first time it is acquired.

At time of writing, we have a bug where singleton mutexes are lazily
initialised, but not initialised in a thread-safe manner. This means
that the first time a mutex it is possible that two threads may hold it
simultaneously.

This test sets up N threads and races them acquiring a singleton mutex
While the test is safe to run multiple times, it can only exercise the
static initialisation code path once per boot.

Original-Bug: 78920
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/544202
Original-Revision: 2bfb579f72abe8ee8baa7247e04a273f97660767
GitOrigin-RevId: f6df88e66e7e5d73cc6519f6bbdf4806fd4a89b3
Change-Id: I7313bf5a57c0629b4fdecd752a6fa731d3fd0c3c