[roll] Roll fuchsia [input] Validate the well-formedness of KeyEvent
The commit at fxr/573576 started enforcing timestamps on `KeyEvent`s
forwarded by the IME service to clients. This was an improvement over
prior state of the art but didn't go far enough.

This change adds verification on the fields `type_`, `key` and
`key_meaning`.  If we forward `KeyEvents` that don't have them this is a
bug, so we crash with an appropriate error message.  I don't think it is
possible to make this check happen at compile time.

Original-Bug: 83632
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/575246
Original-Revision: d665cccfc02406f437d742adbc0455eb9e6366f1
GitOrigin-RevId: acc8ee9a2ae95f484da2c6657688d681cfe34345
Change-Id: I490f972ac15c4167cbaf8d059be4a2170bd28c16