[roll] Roll fuchsia [setui][tests] Allow finer controlled trigger
This CL allows the light sensor scanning code to use a generic trigger
for starting processing of the next light sensor event. It moves the
trigger creation to the caller, so the production code can still trigger
a time based event, but the tests can use a manual trigger to be more
deterministic in the evaluation of the code. This should hopefully
prevent test flakes due to timeouts we've been seeing.

Original-Fixed: 82610
Test: fx test -o setui_service_tests
MULTIPLY: setui_service_tests:10
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/573421
Original-Revision: a81edf463a4c3815bf5e974668319cbba42bd8b8
GitOrigin-RevId: f2dd02125fbe5a8565176f1cfbf07582bba11244
Change-Id: I35b23fe753a59bdb27253beee6401be10807b4ed