[roll] Roll fuchsia [inspect][contrib] Workaround rustc issue
This CL updates the syntax in a macro call to avoid a bug in the rust
compiler that prevents us from analyzing nightly rolls in the rust
compiler. The issue itself is actively being worked on at
https://github.com/rust-lang/rust/issues/88307. While normally we would
not add a workaround for such issues, it's become too complicated to revert
the original change due to commits applied on top of the original
change. We've also been unable to analyze any other potential failures
that might be coming in. Given the active work on the issue, we felt it
was safe to add in this workaround so we could track any other possible
issues coming in during this time.

Original-Fixed: 83801
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/575604
Original-Revision: 1ce17cf0ee852bb093df5d102ea3c366cc0ffe32
GitOrigin-RevId: 734b2c8e9fc357f869d124879f97ef0f0098f7fd
Change-Id: I226cf88e439679acb44127686d3e4c4da1f47c3c