[roll] Roll fuchsia [test_manager] Map Suite epitaph to LaunchError
This CL changes test_manager to read and map the epitaph
of the fuchsia.test.Suite channel in the event of failure
to invoke a method. The mapping is:

zx::Status::NOT_FOUND => LaunchError::InstanceCannotResolve,
zx::Status::UNAVAILABLE => LaunchError::CaseEnumeration.

This change was made in order to prepare for the migration
of ScopedInstance to use `fuchsia.sys2.Realm/OpenExposedDir`
internally as opposed to `fuchsia.sys2.Realm/BindChild`.
This struct is used by RealmBuilder which test_manager uses
to create a hermetic environment for test suites. test_manager
currently maps failures during Realm construction to
LaunchError::InstanceCannotResolve. An example of a failure
it may encounter is that the test URL provided by the user
may be invalid (read: unresolvable). After the migration
to `/OpenExposedDir`, such failures will no longer show up
during Realm construction because this method 1) is async
and 2) doesn't automatically start the realm as `/BindChild` does.

So, in order to remain at parity with the future migration, we
read the epitaph of the channel of the `fuchsia.test.Suite` connection.
This will provide slightly more granular reasoning as to why connecting
to the protocol failed.

For more context on this migration, see
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0108_component_binder_protocol

Original-Bug: 81401
Original-Bug: 81400
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/573463
Original-Revision: 770a9a9f705aafcc815a59216a2b6631e3e2dcdd
GitOrigin-RevId: 69c7ce4fa98b2faaf5a895156c50b1cc2c6b70b3
Change-Id: I0c31ae1d9d59feb1a0015d8499a29c0b76d8973d