[roll] Roll fuchsia [modular] Fix use-after-free in StoryControllerImpl
StoryStorage may outlive StoryControllerImpl because it's owned by
SessionStorage and not destroyed together with StoryControllerImpl.

It's possible for StoryStorage to be updated and invoke this listener
after the StoryControllerImpl is destroyed, for example when an
AddModCall is processed during teardown.

Original-Fixed: 83815
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/575746
Original-Revision: 2e69369bb0eb0ce5f603e7abd6f28912a1409ddc
GitOrigin-RevId: 70f0064a62853d5b6ff175d0522dce83adafe070
Change-Id: I9776a9327ae2545d5499b29d3fbf8aae01ccdfd2