[roll] Roll fuchsia [kernel][thread] Change assert in SetMigrateFnLocked
When a VCPU is being destroyed, we need to be able to clear the
migration function, otherwise the scheduler may call the migration
function and incorrectly refer to freed memory.

It should be safe to set the migration function to nullptr, even if we
are in the midst of migrating the thread.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/574621
Original-Revision: 8a3fa3583ce6d56c3602235f44af8826a3012ec9
GitOrigin-RevId: bc48e3545ba6e15569332a7e4b079a1277db8e69
Change-Id: I4c3ff37d81f296731525013f6ab3f66c1bf00477