[roll] Roll fuchsia [cm_config] Add builtin_capabilities field to RuntimeConfig
Currently, component manager will route any capability that
claims to be a built-in capability. (It only then opens the
capability if it's actually implemented as a built-in.) In
order to statically verify capability routes, we need some
way to specify the intended set of built-ins in a config or
manifest file. This change adds a Vec<CapabilityDecl>-valued
field to the runtime config.

In a follow-up, component manager's routing code (and the
static routing analyzer) will reject any uses of
capabilities that are routed as built-ins but not registered
in this config field.

Original-Bug: 83626

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/574148
Original-Revision: d05cdcdb27cebac6242e03c2149d5e4890492271
GitOrigin-RevId: f98d6083d17b9b3d8f8146e053353dc407c58619
Change-Id: I0a0a9c362d4ecd2a1393e9eb9d43a90ba4d397fb