[roll] Roll fuchsia [usb-virtual-bus-launcher] Fix so that zxtest assertion failures aren't ignored
Before this change, if I add EXPECT_TRUE(0) somewhere in
USBVirtualBusBase() in
zircon/system/ulib/usb-virtual-bus-launcher/usb-virtual-bus-launcher.cc
(or add ASSERT_TRUE(0) at the end of that function), an assertion
failure message gets printed in the test output, but the test passes
overall.

That happens because this code is linked into a dynamic library and
links against zxtest, so it gets a separate copy of the zxtest library
(which is currently statically linked) and hence has a different
instance of the global variable that tracks whether an assertion
failure has occurred.  The copy of zxtest used by main() therefore
doesn't see the failure.  zxtest currently isn't safe to be used from
dynamic libraries this way.

Fix this by statically linking this code.

Original-Bug: 83770
Test: see above
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/575792
Original-Revision: 0d5919edc4ced4eb746317f901837e4d7658546a
GitOrigin-RevId: de2f35a80e6bdf6065175964b6c272481acf1a1e
Change-Id: I78acbece1b0a7366cd61ff97847c0e1eded64f1b