[roll] Roll fuchsia [fidlc] Add attribute for overriding anonymous layout name
See bug for context. This attribute is placed on the layout
directly rather than on the member, since it does not
affect the anonymous naming scope (introduced by the
member name), just the generated name itself. This allows the
result union special name handling to be reimplemented
in terms of the @generated_name attribute.

Implementation is straightforward but for one aspect:
the override name is required for compilation, but we
don't have access to the resolved Constant value of
the attribute arg before compilation. To get around this
we use the span of the attribute arg itself, and do
validation of the argument/compilation later. This also
means that Attribute::GetArg returns a Constant rather
than a ConstantValue, so that we have access to the
attribute arg's span.

Test: fx ninja -C out/default host_x64/fidl-compiler &&
./out/default/host_x64/fidl-compiler

Original-Fixed: 81550
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/568581
Original-Revision: aa1656d16326cd20325e981e60b0dd270048adb0
GitOrigin-RevId: 0ec933e16a713db0435ae8348aa116eb759df452
Change-Id: I9d36e9c92aaa9988ecc3c92ae9e73b162cb9b4fc