[roll] Roll fuchsia [cm][runtime_config] Register builtin capabilities
In a follow-up, component manager's router will reject
uses of capabilities that are routed as built-ins but
are not registered in builtin_capabilities field of the
runtime config.

This change registers the builtin capabilities that
are used in real and integration-test instances of
component framework.

Original-Bug: 83626

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/556667
Original-Revision: 313713b588e7be620c8fa7a457d0e783accde2b6
GitOrigin-RevId: 40fc2b52acab7906624c78d09c70eeb5ad1cf048
Change-Id: Ie88109c82eb0717247a13b1f389cb3066218ca73