[roll] Roll fuchsia [identity] Add Identity API methods to support authn prototype.
This is the second in a chain of CLs to adapt the existing identity API
surface to the simple authentication prototype that we're in the process
of building.

Step 2 is to add in temporary methods (marked as deprecated) to get and
to create an account under the hardcoded assumption that authentication
is password. This also adds an additional longer term method to get
account metadata from the account manager interface.

Original-Bug: 81605
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/573578
Original-Revision: eb2b92e34c45e75b70c7fe5bf3768b7b8cf40058
GitOrigin-RevId: 96cab275f33600bef42e96db90784a6250e8d520
Change-Id: I9cd5fce37ca193b71ddf1071867ae266135f3946