[roll] Roll infra/recipes [symbolize] Return stdout in case of timeout
Otherwise the function returns None. Using None as the contents of a log
is invalid and causes recipe engine crashes. For example:
https://ci.chromium.org/b/8837286647449398209

A symbolizer timeout is no longer considered fatal, so just return as
much stdout as the symbolizer was able to produce before the timeout.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/575857
Original-Revision: 0843adf09ba5ecb3772d02153648385b8aaf2189
GitOrigin-RevId: 84c104b38f14bbafc8a4c75b239d79308f4ac7d4
Change-Id: I1ee4152137213ebb585b93a1b73e0e9fe9ad918e