[roll] Roll fuchsia [connectivity][fuzzing] Remove zxtest dep from usb-cdc-ecm fuzzer.
The usb-cdc-ecm fuzzer had a GN dep on zxtest. It is unneeded, but
didn't harm anything for fuzzers running as v1 component. For v2
components, it introduces a dependency on the elf test runner that
conflicts with the generated fuzzer unit test's dependency on the gtest
test runner.

Original-Bug: 76528
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/573572
Original-Revision: ad46f137309a46928274eca4df9656cc4cfe6ddf
GitOrigin-RevId: 2ca9aceaeb1cb856c54b5a160df665ce5d101e51
Change-Id: Ib466286f2b00012d54b815f4280d3e92e9ba010a