[roll] Roll fuchsia [ftln] Remove unused data scan in spare validity calculation.
The original intention was to check for anything that appears to be
non-erase bytes in the data section of the ECC page as well, and that
information was scanned for but erroneously unused. Experience so far is
that we have always had partial writes on the full page boundary, but
even in the case of partial writes on the ECC page boundary this should
still catch the issue since the spare bytes are spread out across the
ECC pages.

Original-Bug: 83474
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/575748
Original-Revision: 8b5da64b6bba8ad5b06f451732bfcf789984feed
GitOrigin-RevId: 27bdbc09c99c335e1185293ae1de225c6faf72be
Change-Id: I0d13665049209521ed0549802a833d95d8305d46