[roll] Roll fuchsia [phys] Clarify semantics of BootZbi::Load()
It is 'morally' the case today that the associated kernel image of a
BootZbi is loadable in-place after a successful Load(); however, the
class's internal bookkeeping does not reliably recognize that. This CL
fixes that, begins enforcing the invariant with a debug assert in
Load(), and simplifies physboot's interaction with BootZbi in relying on
that invariant. That simplification of physboot also removes the buggy
assertions and expectations that led to fxbug.dev/81726.

Original-Bug: 68585
Original-Fixed: 81726
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/574782
Original-Revision: 95c10c5cc3c1bebc92db9412ca3a9845357519bf
GitOrigin-RevId: 103594d295d187de0ffbf82fae664758d0f50951
Change-Id: Ia5b73ff43a91713e7467579ef0a0c8febefa29c3