[roll] Roll infra/recipes [fxt] Always categorize `launch` failures as reds
If the launch step fails, it is either an indicator of real failure to
do with the CL, or is an indicator of infrastructure failure in g3.
Either way, this is not attributable to Fuchsia's infrastructure and so
should not be purple.

Original-Bug: 83834
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/576411
Original-Revision: 3f44372a6043ed2aae1bab1208a19b65ed24432d
GitOrigin-RevId: e83fba879cb3511b4d4d312727042602f2663d2c
Change-Id: I1163baa673c3258ecf98dd1df98764a95a53b2ea