[roll] Roll fuchsia [brightness] Remove frequent service startup logging
Workstation shell connects and disconnects from various settings related
services when displaying sysui overlays. This is to prevent background
load impose by sysui when it is not visible. Every time the sysui is
displayed, it connects to the settings services, in this case brightness
service, and cause this log to be displayed.

In a typical usage, for example in workstation build walkthroughs, the
sysui will be displayed tens of times within a minute. Having this log
is not very helpful and not consist with other setting services.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/576221
Original-Revision: 0f821e5b7178e9fdd7f445a7211d57fd0e814d0f
GitOrigin-RevId: e6e66bfb6b8a9c349e27438b80879b65931d462e
Change-Id: Iea5d881880215ad3876d90ec2cc013c4e6119b80