[roll] Roll fuchsia [display][i915] Prevent client GTT allocations over framebuffer
The bootloader framebuffer is located at the base of the "data stolen
memory" region. When initializing the address mapping in the global GTT,
the driver initializes the region allocator in a way that allows clients
of fuchsia.hardware.intelgpucore to modify the contents of the
framebuffer directly.

This is now prevented by excluding this region from the GTT region
allocator.

Original-Bug: 80350
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/566901
Original-Revision: a330ddf92ede9e9bc6bd0df332131d4d7f35d3f8
GitOrigin-RevId: 3a495022e17f40cad90c2c5a4861721454c8e01e
Change-Id: Idd47a5efb8cf8ef7ea39bcc52e35ee03bc704f99