[roll] Roll fuchsia [wlan] Explicitly parse and handle ie::SupportedRates
The ie::SupportedRates newtype should be used instead of u8 when a
rate is not specifically used in a FIDL struct. In particular, the
parse_*_rates() functions are now used to ensure the parsing logic is
used when parsing the Supported Rates and Extended Supported Rates
IEs.

Test: fx test //src/connectivity/wlan
Original-Fixed: 83639
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/574558
Original-Revision: 70906443f35351e6a0df8e225e9e1e83919e1ab9
GitOrigin-RevId: 180c6058057126065690865847ceccc610dbeb3d
Change-Id: I7de8909449211c8cee1817d668dc1370373a771b