[roll] Roll fuchsia [pwrbtn-monitor] Support setting action to perform on power button.
On Chromebooks, the power button is used by the TPM to verify that a
user is present at the device. This is necessary when enabling
case-closed debugging or when using the TPM as a U2F key. Before this
CL, pressing the power button in these scenarios would shut down the
system.

This CL adds an interface that allows components to configure the
behaviour of powerbtn-monitor so that the system isn't shut down in
scenarios like this.

Original-Bug: 82504
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/573203
Original-Revision: 4e7a00187bbaf0e10c5925c6e1c864e1b0923120
GitOrigin-RevId: 11d869d3e6e65752314f5a15c95d081e6c31367b
Change-Id: I22f70162f0a321236ece44500f1d8c02035a6624