[roll] Roll fuchsia [realm-builder] place services in the v1 realm
Provide v1 components launched by realm builder with their services by
providing them in the `additional_services` field of the singleton realm
that is constructed instead of by providing them to the component
itself. This means that any other v1 components that are manually
launched in that realm will be able to access these services as well,
which is more in line with common expectations in v1.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/576102
Original-Revision: 60c0d0877038caacff895c89a4fffe61a0453596
GitOrigin-RevId: 3b03e304cc51528fe67114f1b8caac0854c58b3b
Change-Id: Id38f1cf97c2d8d8ca08b7a1ef36bd90a9f39b72c