[roll] Roll fuchsia [safemath] Replace zx_library() with library_headers().
The safemath header-only library can be modeled by a library_headers()
target instead of a zx_library() one. This avoids the creation of an
empty libsafemath.a file that is linked into all dependents, without
changing the semantics of the target.

This is the only use of 'kernel = false' in a zx_library() instance,
which is replaced here with an explicit assert. This CL thus prepares
for the removal of the 'kernel' parameter from zx_library() in a
future CL.

Original-Bug: 83881
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/576425
Original-Revision: dfb9be109e44caa6310f56aa276cb2b9343ee620
GitOrigin-RevId: ff6cdccfe40cee7d1b84468a61700f4a77754dda
Change-Id: I96f55a456b6237a3cc4babf1661e097ea32c5f97