[roll] Roll fuchsia [codec_factory] Initial migration to v2
This change modifies Codec Factory so that it can run as a v1 or
a v2 component. The code looks at what services are available in
its environment to decide whether it is running as a v1 or v2
component. If running as a v1 component the Launcher service is
used. If running as v2 component the Realm service is used.

This change adds v2 test components which use the same code used
in v1 which validate operation of the software-based codecs.
Future changes will add more v2 tests.

This change does *not* move Codec Factory to v2 in production.

Original-Bug: 78647

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/563506
Original-Revision: 905763f2c143e60ac224d15ebd526cb0f4d80642
GitOrigin-RevId: be304a396dad3a45f1ad97ca60e8647416872467
Change-Id: I4719072d78ba95a31342ffd5d20a0462be8e7231