[roll] Roll fuchsia [llcpp] Use move semantics for LLCPP resource unions
After the move to the v2 wire format, union values with payloads <= 4
bytes will be inlined on to the union. For the case of handles, this
means that the handle value will be stored on the union rather than
having a pointer to the handle.

A consequence of this is that unions with single handles (sent alone
or in wrapping structs / arrays) will no longer be able to be
copy constructed or copy assigned. For the sake of providing a
consistent API, this CL makes all resource unions move-only.

Original-Bug: 79578
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/577261
Original-Revision: 1bfc30eff4b777a7309d3585c2e2164593669d2d
GitOrigin-RevId: 7e6dc48f7af79d7155e91ce7fb6c813ce32f1d8c
Change-Id: I836760ecb0140319ae35c04acd56ff23621a4402