[roll] Roll fuchsia [identity] Remove account transfer functionality
A couple of years ago we protoyped functions to transfer Fuchsia
accounts between devices using Overnet. The notion of a platform account
that spans devices is no longer in scope for Fuchsia so we should remove
this and the associated complexity from the API surface.

Account transfer was the only feature to use the "prototype mode" of
building account manager with features that weren't ready for prime
time, and structured config would be a better fit to meet this need if
it arises again in the future so this CL also removing the notion of
the account manager prototype.

This was pretty cool and I'm sad to see it go, but simpler is better.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/576534
Original-Revision: e5e08a0d2b5fd8cc2838c666c1f0ffb1d2c4e913
GitOrigin-RevId: 6eaa3ece0a474c8da4ab5d28b5b55dde3395e626
Change-Id: I25e469d5e3dca7128058167653b103f38db2e1e6