[roll] Roll fuchsia [scrutiny] Hard fail on collection failures.
Since Scrutiny has moved to an in-memory data store for the model hard
faling on collectors makes more sense. Prior to this change if a
collector failed to collect required information Scrutiny would still
launch and work with an incomplete model. This was acceptable for when
the framework was used only as an auditing tool but as its scope has
expanded to being a compiled_action failures should occur as early in
the pipeline as possible. To achieve this I added a new
CollectorState::Failed which allows errors to be propagated and to abort
the running of scheduled collectors. This should greatly improve the
debugability of the framework.

Original-Bug: 81806
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564130
Original-Revision: 9065b0755f80f3c48bdce11ad3339e72665eeb07
GitOrigin-RevId: ab0c9ef0da462abd4bc06c0eb7f2d3d4cdf4c31a
Change-Id: Ie736f5a9e57609f6a3abb8b1767d8ac3a3596dd8