[roll] Roll fuchsia [lsc] Start realms via fuchsia.component.Binder
This CL proactively migrates certain realms constructed via
RealmBuilder to connect to the Binder protocol of the Realm's
root. It only does so for cases where no other connection is
made to the realm's children. This is done to prepare for the
migration of `fuchsia.sys2.Realm/BindChild` to `fuchsia.sys2.Realm/OpenExposedDir`
in ScopedInstance, which RealmBuilder uses internally. Currently,
after a realm  is created via RealmBuilder, it is automatically started.
However, after the migration, this will no longer be the case.
As such, tests relying on this behavior will break.

For more context on this migration, see
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0108_component_binder_protocol

Original-Bug: 81401
Original-Bug: 81400
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/574342
Original-Revision: e24d5ce4bfb9836026b07aadec4883ab02ee695e
GitOrigin-RevId: 122cc09c059ae84937cd1c4f3bcd22663e68d278
Change-Id: I400486e4491c95077daa7e0b4f904505bb46f567