[roll] Roll fuchsia [lsc] Start ScopedInstance via fuchsia.component.Binder
This CL proactively migrates certain components constructed via
ScopedInstance to connect to the Binder protocol of the instance's
exposed directory. It only does so for cases where no other connection is
made to the instance. This is done to prepare for the
migration of `fuchsia.sys2.Realm/BindChild` to `fuchsia.sys2.Realm/OpenExposedDir`
in ScopedInstance. Currently, after a an instance is created via ScopedInstance,
it is automatically started. However, after the migration, this will no
longer be the case. As such, tests relying on this behavior will break.

For more context on this migration, see
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0108_component_binder_protocol

Original-Bug: 81401
Original-Bug: 81400
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/574410
Original-Revision: 406e8f6c4f09b13772ff156a8a9b9f52de0d96df
GitOrigin-RevId: a8812705acc664d42046bf3260262731523a2c29
Change-Id: I17f0b027bf25d60bb6e2e1d2d35fb7ceda388115