[roll] Roll fuchsia [lsc] Start realms via fuchsia.component.Binder
This CL proactively migrates certain realms constructed via
RealmBuilder to connect to the Binder protocol of the Realm's
root. It only does so for cases where no other connection is
made to the realm's children. This is done to prepare for the
migration of `fuchsia.sys2.Realm/BindChild` to `fuchsia.sys2.Realm/OpenExposedDir`
in ScopedInstance, which RealmBuilder uses internally. Currently,
after a realm  is created via RealmBuilder, it is automatically started.
However, after the migration, this will no longer be the case.
As such, tests relying on this behavior will break.

For more context on this migration, see
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0108_component_binder_protocol

Original-Bug: 81401
Original-Bug: 81400
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/574411
Original-Revision: efef51496b2c8a90c6724212c42f6b191dbfef2d
GitOrigin-RevId: b2943f358ca19338d1107682ff8eacb6b2b3d2cc
Change-Id: I55d7af775645a225211f9e84594812e96a0549e7