[roll] Roll fuchsia [lsc] Start realms via fuchsia.component.Binder
This CL proactively migrates certain realms constructed via
RealmBuilder to connect to the Binder protocol of the Realm's
root. It only does so for cases where no other connection is
made to the realm's children. This is done to prepare for the
migration of `fuchsia.sys2.Realm/BindChild` to `fuchsia.sys2.Realm/OpenExposedDir`
in ScopedInstance, which RealmBuilder uses internally. Currently,
after a realm  is created via RealmBuilder, it is automatically started.
However, after the migration, this will no longer be the case.
As such, tests relying on this behavior will break.

For more context on this migration, see
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0108_component_binder_protocol

Original-Bug: 81401
Original-Bug: 81400
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/574412
Original-Revision: bd7e06bcffe7485699517311fff9059d24555147
GitOrigin-RevId: 9828580ffe351bf0e94c3de383d50be5771dc8b7
Change-Id: I566a63360f394622b7974054e66e4274e522079d