[roll] Roll fuchsia [lsc] Start realms via fuchsia.component.Binder
This CL proactively migrates certain realms constructed via
RealmBuilder to connect to the Binder protocol of the Realm's
root. It only does so for cases where no other connection is
made to the realm's children. This is done to prepare for the
migration of `fuchsia.sys2.Realm/BindChild` to `fuchsia.sys2.Realm/OpenExposedDir`
in ScopedInstance, which RealmBuilder uses internally. Currently,
after a realm  is created via RealmBuilder, it is automatically started.
However, after the migration, this will no longer be the case.
As such, tests relying on this behavior will break.

For more context on this migration, see
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0108_component_binder_protocol

Original-Bug: 81401
Original-Bug: 81400
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/574588
Original-Revision: 27095b88b746e15bf53fa2379da595fc20eef129
GitOrigin-RevId: 15dd71a2c30f4411ccf06ed35875d4f7122653d9
Change-Id: I6667363257db8beb1297abd59e69fee63e1e3cdb