[roll] Roll fuchsia [archivist] Block on component Start for test.
This CL changes the log::attribution::log_unattributed_stream
integration test for Archivist to block on start of test realm.
It does so to proactively prepare for a semantics change in the
RealmBuilder library.

A dependency of RealmBuilder, ScopedInstance, will migrate from
using `fuchsia.sys2.Realm/BindChild` to
`fuchsia.sys2.Realm/OpenExposedDir`. The primary consequence of
this change is that the instance created ScopedInstance won't
be automatically started. Since RealmBuilder creates its Realm
with this struct, it too will change.

In order to not break CI/CQ, and to keep this test green, the
test was changed to block on the Start event.

For more context on this migration, see
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0108_component_binder_protocol

Original-Bug: 81401
Original-Bug: 81400
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/574589
Original-Revision: 1c99def3344e0d85b264aca54924993767ffba79
GitOrigin-RevId: 1b165742ccb105ce328a5ed694b6c15dffd01c3b
Change-Id: If46d7d3af7a84fcc82f4803b8e40748bd2049c2f