[roll] Roll fuchsia [component_manager] Migrate ScopedInstance uses.
This CL proactively migrates uses of ScopedInstance in
test to connect to fuchsia.component.Binder of the exposed
directory of the instance. It only does so for cases where
no other connection is made to the instance. This is done
to prepare for the migration of `fuchsia.sys2.Realm/BindChild`
to `fuchsia.sys2.Realm/OpenExposedDir` in ScopedInstance.
Currently, after a child component is created via ScopedInstance,
it is automatically started. However, after the migration, this
will no longer be the case. As such, tests relying on this behavior
will break.

For more context on this migration, see
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0108_component_binder_protocol

Original-Bug: 81401
Original-Bug: 81400

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/573464
Original-Revision: 87146c791fdd0d7ffd60ce7dc4b60a4b200c7b90
GitOrigin-RevId: 979a1bd8e4b95059e036c85063a05f5e0a84f6d6
Change-Id: Ib4818f37bf1ca30d4f667247a45e8fda774d18f1