[roll] Roll fuchsia [kernel][dlog] Fix dlog skipping empty newline
If a record contained only a newline character, it was removed and an
empty record was written to the serial without a newline character. This
CL updates that logic to not skip the only newline logs and to maintain
the contract that all logs end with a newline.

Original-Bug: 83968
Original-Bug: 83566
Test: k ut debuglog_tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/576909
Original-Revision: d0587d35ee209ed542b00ff37eb5c1bcaefa73ff
GitOrigin-RevId: 0fa14fe1544b3befe97afa69a68e82a461f9e549
Change-Id: Ic1c9fc7534d297744ddcfd351bcdb59990f49b78