[roll] Roll fuchsia [netemul] Truncate endpoint names to fit ethertap
Truncate endpoint names by dropping characters from the start of
the string to fit the maximum limit imposed by ethertap, which is 30
characters long. Characters are dropped from the front because the only
names likely to exceed the length limit are names of tests, and taking
a suffix results in a more descriptive string since nesting of
testcases appends suffixes. Note that there is no limit for netdevice
because it doesn't have a name.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/576908
Original-Revision: c673ac600590b32645a2a8e127e88885c7c2bef5
GitOrigin-RevId: e6e06a7d1a5fc92f0f92a828c2dd47c332da78a4
Change-Id: I70935f0fbf49044edeca4b9d2317734c0f9d54dd