[roll] Roll fuchsia [identity] Delete identity key manager.
Identity key manager was a (not yet fully implemented) way to manage a
consistent set of cryptographic keys across all the devices a Fuchsia
account was provisioned on.

Platform level portable accounts are now out of scope for Fuchsia and so
this work in progress should be removed from the API surface.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/577063
Original-Revision: 3484df1dea92317b61532d8145222292c3d9989e
GitOrigin-RevId: f77ec71d478fcc5f01ca6697ac8b5e0ad0ff6f16
Change-Id: Ie6d2a83225d0876ea6bfd2e5d2be65c51431e29a