[roll] Roll fuchsia [llcpp][fake-drivers] Copy radiobound bytes to std::vector
Currently, the fake-ot-radio test pushes a VectorView to the
radiobound_queue, but the lifetime of the VectorView's contents is only
up until the end of SendFrame(). This leads to a potential
use-after-free bug. Copying the bytes to an std::vector avoids this
issue.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/578121
Original-Revision: 523da2a17c4a4bb4fcb40e1fb89260d1b11d38a1
GitOrigin-RevId: 71c3d117e6adec92a126d236906a16a55ddbfe0c
Change-Id: I7082d8f21c51f5fa2600b5af07c31bb5090e39c6