[roll] Roll fuchsia [fvdl] Add dry-run flag and serde to ffx emu
In preparation for adding FVD metadata to ffx emu, we need to be able to
use a JSON file as input and quickly observe the effeects of that file
on the command's output.

This CL adds serde (rust (de)serialization to/from json) to the plugin's
args structure, and a dry-run flag to terminate the plugin as soon as
the command line is formed. This enables quick iteration times as we
work on defining the metadata schema for FVD going forward.

Original-Bug: fxbug.dev/70828
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/577421
Original-Revision: f3382ad1f77bbe721ae0075e58e3884e5f5cd18a
GitOrigin-RevId: c3e53073b0ea5247dc414bc9e8eac6912577a4b9
Change-Id: I980389c89304ad3e23b4ede9d45c028ffcc98c36