[roll] Roll fuchsia [virtualization] Modify termina_guest kernel/CIPD scripts
Update the "update_cipd_kernel_package.sh" script (which builds and
uploads a kernel to CIPD) to be better suited for termina_guest instead
of linux_guest.

The main semantic difference here is that we no longer attempt to
package up userspace images in the CIPD package, but only put the kernel
in.

Additionally, we update the CIPD target to avoid using the "linux_guest"
name.

Original-Bug: 84205

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/577630
Original-Revision: 14f084523158bfec7aa51b901946ef7720cfd45f
GitOrigin-RevId: e8059e47b7885056fddf28735f41188b3f9cb1bc
Change-Id: Idea7ec4d27fc8b8b84cabee1ae331a2614656045