[roll] Roll fuchsia [kernel][vm] Pass PageRequest in MapRange
Technically if the SW_FAULT flag is set a PageRequest must be provided
to LookupPagesLocked, although the caller can avoid this if it knows
the target VMO is not pager backed. As such whilst this is not
presently an actual bug, it costs nothing to pass in a PageRequest and
will allow graceful error detection in the future should the current
assumptions get violated.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/574810
Original-Revision: 6f0276e58445f5d8e29d728e8ea317e6a6762b79
GitOrigin-RevId: 0f3ff60d7f09391571e9a1e099a7e7569adb1869
Change-Id: If523d76cb985bd04de11fde9d8695fd83b002c34