[roll] Roll infra/recipes [coverage] Add property to collect absolute coverage on tryjobs.
Before, we assumed the fuchsia-coverage tryjob would only be run on a
gerrit change to collect coverage for the changed files on the CL so we
would skip launching builders or instrumenting files that wouldn't
contribute to the coverage data for the changed files. However, it is
worth having a tryjob that can run the fully instrumented build and
collect absolute coverage so that we can test that changes such as
toolchain rolls don't break anything in the coverage pipeline.

Original-Bug: 84219
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/578361
Original-Revision: 09f59ae014945636a5310937b65562dddec3b2f1
GitOrigin-RevId: 00aa7f0ec9322f52ca09128ff991a170cd6a5bd3
Change-Id: I20d50f109d89edca678ccc0e59c9dc38e4f41397