[roll] Roll infra/recipes [build] Add option to produce codesize report
Creating a codesize report from the `codesize` tool is gated by a flag
for now while we incrementally fix the integration issues. When it is
working, we'll merge it into the size_checker output and run it under
the same conditions as we run size_checker.

Original-Bug: 84202
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/578421
Original-Revision: d194817ce042b47b1dd711be1dfa9e4b1ad59d1d
GitOrigin-RevId: 95c4e776719630fb54b57acb40c1028c3f74faf1
Change-Id: Ia738a6f62332eb80fdbbff4bf7760e8af642f76b