[roll] Roll fuchsia [core][test] Rotate queues more times for eviction.
Currently there is a single active queue, and so one rotation is
sufficient to move all pages into the evictable state. However, the
number of active queues is changeable and not visible to this test.
To attempt to future proof this test, make it rotate a few times.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/574807
Original-Revision: e7dd2274c556049a2830d8d29bc55f982b09a36f
GitOrigin-RevId: 3c5e97d719eb8e918c7b371510ed91fb31271c84
Change-Id: I0fed06be44c0bb604eda4dca8cb04c907581e47d