[roll] Roll fuchsia [vulkan] Prototype loading validation layers from package
The validation layers are now included in a component that just serves
them as a directory capability. Applications that want to use the
validation layers can launch layer-server as a component and use the
directory capability. To make everything is hermetic as possible, the
client application can use a specific package hash when launching the
server, ensuring that the validation layer version is the exact same is
the client was compiled against.

Original-Bug: 83307

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/574025
Original-Revision: 3963682e657ab6240785a4f756429d777fa7af6e
GitOrigin-RevId: 0e1db905f15510fff32363cfe216d776dc0c59cb
Change-Id: I0a851b5930661dd0ca43f22d2fa503aa6a8a3fd5