[roll] Roll fuchsia [archivist] Change test to use exposed dir
This CL changes the `log_unattributed_stream` test
to connect to LogSink via the exposed directory of
the root of the test Realm instead of its hub path.
This is done because the latter path was causing flaky
behavior for reasons that have yet to be understood.

Original-Bug: 81401
Original-Bug: 81400

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/578861
Original-Revision: 5080a433ce24f36b42fa57d03084fe30099c44c4
GitOrigin-RevId: cc0e2169d4b341c44f76578366017bb80c815c22
Change-Id: I5fdb1c3219eab615f8561eab69b79b4809393f84