[roll] Roll fuchsia [fxfs] Break up flushes into several transactions
Breaks up flushing into batches of 1MiB of data at a time.  This is
necessary to avoid overwhelming the pool of transfer buffers, and may
also make flushing large amounts of data more reliable since we have
less risk of exceeding the transaction size limit.

The file size is updated only as far as the data has been flushed, so
that if we interrupt flushing, the file size and the data written so far
will be consistent with each other.

This required reworking the logic for reserving bytes for the
transaction metadata slightly: We now have multiple transactions per
flush call, which means we need multiple reservations for the metadata
of all transactions.  We accomplish this by allocating more metadata
space each time we dirty the next 1MiB of data.  The writeback cache no
longer keeps track explicitly of how many bytes it has reserved; this is
tied to the number of dirty bytes in the cache.

Also fixed two small bug:
- Pending crtime/mtime updates would be dropped if a flush failed
  (rather than going back into the cache).
- In some cases, ReadResult::MissingRanges could request more than
  READ_AHEAD_SIZE bytes in a range, which breaks an invariant relied on
  by the caller.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/574102
Original-Revision: 6d053c8a2edb86fac12dbc3c29f77e554ada5d95
GitOrigin-RevId: 979748480d7e54423fe72302b11a9353b634b4b1
Change-Id: Ie1690b868fe750304af098c91cff5862e361810d