[roll] Roll fuchsia [kernel][pager] Make PageProvider RefCounted
Changes the PageProvider reference in the PageSource to be a RefPtr,
instead of a unique_ptr. For this commit this is effectively a no-op,
but it unlocks subsequent changes that need additional locations to
hold a reference to the PageProvider.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/577626
Original-Revision: e16b43578b71619a0b9be7625a537644435eb65b
GitOrigin-RevId: d9a0e7464ab48ffac1a8c4e84bfbad8d3d39c511
Change-Id: Iec306d86a177f7c4045b12c3150b2ca47d2645c9