[roll] Roll fuchsia [cm] Allow dynamic children as sources/targets.
In cm_rust, `OfferSource` and `OfferTargets` s can only be static
children, not dynamic children in collections. This means most of the
component manager code base can't even model a dynamic offer. This
change makes it so `OfferSources` and `OfferTargets` can have a
Collection.

I don't love this solution. It introduces many places in the code
where an offer is known to have a static child (i.e., any offers that
are defined in ComponentDecls), but the type system now forgets that
fact. It'd be nice if there was a separate set of types for
statically-declared offers as opposed to offers as they exist at
runtime. Unfortunately, `OfferDecl` is used to mean both a static
offer and a runtime offer throughout the CF codebase, and changing
that would be a much larger project.

This CL aims to be as "yadda yadda" as possible. Dynamic children are
still not supported, and if a ChildRef is found with `collection` set,
the code will panic. Actually addressing the dynamic cases will come
in further CLs.

Original-Bug: 81207

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/544381
Original-Revision: 1e39f8884c69d405cd4c2a9bf91463c22a3f3ef8
GitOrigin-RevId: 5adc09092b8ffe674b05481675b1250c4184d1e2
Change-Id: I8604acf38a1bd4858f6a7c8a65b8e799b1998519