[roll] Roll fuchsia [flutter] Fix runner crash when switching between views
This change fixes the crash observed in the flutter runner when the user
switches views on the Workstation shell. This is because the view was
destroyed in FuchsiaViewController.dispose, when the PlatformView
was taken Offstage. This is flutter behavior when a PlatformView is
moved offstage when switching. Calling view destroy here is wrong because the
underlying scenic view and the owning component/app is still alive and
the user is expected to switch back to the app.

There is no use case to call destroy view method on the
FuchsiaViewsService. Rather, the runner notifies the client code that
the view was destroyed in response to the underlying scenic session
closing. The method should be removed to reduce developer confusion.

Original-Fixed: 83841
Test: fx test --host fuchsia_scenic_flutter_unittests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/579147
Original-Revision: 6ad52ef6267c07b90f61c5de7d526d3ca79e38bd
GitOrigin-RevId: 0a14a549b9f6343b5a77b2bed534acc7615f28f4
Change-Id: I8d5837a54d0f21f3a328dba7270cc492cfb165d0