[roll] Roll fuchsia [fx][scripts][fish] Fix completions when there is no vendor/ directory
When cloning the repository, there is no vendor/ directory. As per
https://fishshell.com/docs/current/language.html?highlight=exceptions#wildcards-globbing,
this leads to an error when expanding the glob $FUCHSIA_DIR/vendor/*/scripts/devshell/.

This change uses `set` for the glob, which does not lead to errors if
the result is empty.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/578942
Original-Revision: 5c3cf3b812ea68f7c8c9016d8c539a8cd304e88c
GitOrigin-RevId: a5cdafcda9a8a4b95a863160c6d2e99ed4bf92f6
Change-Id: Ie3541a844489007b4a8c0643d7a72408e8669036