[roll] Roll fuchsia [starnix] Update socket to read from pipe
Previously the socket would write to the pipe that it owned, but this
change flips that around so that a socket reads from the pipe that it
owns and writes to the pipe of the connected socket.

This makes the read/write file logic a bit more sensible, since we can
read out remaining bytes even after the connected socket has been
disconnected.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/579961
Original-Revision: 84abf791edc8ab71831ec5e6ede8b8e661f7aac1
GitOrigin-RevId: 4bfc2de2ef9b0d84e02ea6cfac678cc7794fa457
Change-Id: If7aac4ab053dbb2a0b633676ec2d2a5167e459bc