[roll] Roll fuchsia [wlan][iwlwifi] remove iwl_mvm_update_quotas_single_iface()
The TIME_QUOTA_CMD crashed the firmware while disassociation
on 8265 and 9260. However, this command is not mandatory.
Since this function only sets the default value in the single
interface situation (the only mode we currently support), we
can safely remove it. A TODO(43218) is already filed to support
multiple interfaces.

Original-Bug: 81770
Tested: on 8265 and 9260, 100 * (assoc + disassoc)

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/573481
Original-Revision: 7f74167cb68cb07c6bdcc9a463c34eae74110f43
GitOrigin-RevId: 5f152b6f9e1ffe541ee0a176762554e111678c5c
Change-Id: I974b22dd7dfd15a8b8f7ce6cdf5d2f221463791b