[roll] Roll fuchsia [bt][common] Assert on impossible AD parsing error
bt-host::common::AdvertisingData can be parsed from raw bytes in TLV
format. The raw format's length field is a uint8_t. As such, fields
which represent lists of objects have upper limits on the number of
objects (based on how many of an object fit into uint8_t::MAX).
AdvertisingData enforces this limit for service UUIDs in AddServiceUUID
by checking that the current amount of service UUIDs is not at the per-
UUID-type limit (i.e. AddServiceUUID method is fallible), because
AdvertisingData may be populated from places other than raw bytes.

However, because these limits are defined based on the raw byte
representation of an AdvertisingData, it is impossible to exceed them
when parsing an AdvertisingData from bytes. As such, when parsing an
AdvertisingData from bytes, we now ASSERT that adding the service UUID
was successful instead of propagating the potential error.

Original-Bug: 79621
Test: Added AdvertisingDataTest::MaxOutServiceUuids to verify that
      adding the maximum number of one type of Service UUIDs does not
      cause assertions.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/579904
Original-Revision: 20bfe29bc500075bcca1c151b964367bb314b39e
GitOrigin-RevId: 62d05d29eafa5893338c7a8fe8552347256a4c72
Change-Id: I0a08bb7446706a5eecd163b4a7434f993c2fc807