[roll] Roll fuchsia [audio][tests] Specify Impulse channels
With this CL, the HermeticImpulseTest::TestCase can optionally
specify which channels should be validated for the presence of
the expected impulse. Previously, we would expect this from every
output channel, and this remains the default behavior, if the
"channels_to_test" field is unspecified.

This is part one of a two-repo change, to re-enable pipeline
Impulse testing on Sherlock.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/579906
Original-Revision: f71b3d54f06bced8bb6c298908f23110cf3824fd
GitOrigin-RevId: 3450e13bf668c6fdbcfd63d3f69b5e3be8fa26ee
Change-Id: I3a774778fcdf43f791a7ecf5bfe312ccbfe1fb55