[roll] Roll fuchsia [fidlc] Split up new_syntax_tests.cc
During the syntax migration work, it was convenient to have a file
to dump tests that applied specifically to the new syntax. It no
longer makes sense to have this file, since it makes it harder to
know where a feature is tested. This CL splits up the test cases
from this file into the appropriate feature-based test files,
ensuring that no test coverage is lost in the process.

A large number of tests in this file exercised the layout parameter
and type constraint code which is new to the new syntax; these
tests have all been moved to types_tests.cc since they relate
to type compilation

Test: fx ninja -C out/default host_x64/fidl-compiler &&
./out/default/host_x64/fidl-compiler

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/577561
Original-Revision: 4b85f0b9cc5edcc2ff07b3d92263ba7901c4ec79
GitOrigin-RevId: 2f26d587ac9d46e4e13467b9e29d99bb87aa8562
Change-Id: I1ed6a3ccc3b11505ac77081a5778950c2942c5da