[roll] Roll fuchsia [bt][gatt] Handle service changed notifications
Add `removed` and `modified` parameters to
gatt::RemoteServiceManager::ServiceWatcher.

Handle Service Changed notifications in gatt::RemoteServiceManager by
performing service discovery, calculating the difference between
current and new services in the changed range, and notifying the
ServiceWatcher of the service changes.

Service Changed notifications are queued for processing. If a
notification is received during RemoteServiceManager initialization, the
notification is queued and processed as the last step of initialization.

When a service is modified, destroy the old RemoteService and replace
with a new one in order to clean up ongoing procedures and stale
characteristic state. This will also alert clients that the old service
has been invalidated.

Original-Bug: 71986
Test: fx test bt-host-gatt-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/570082
Original-Revision: 6152fccac9184d61c0576b10633c87cef6591beb
GitOrigin-RevId: 11bc99e37e4b3bfe13d18f53bb98c78af87986ae
Change-Id: I1e15874a022ffa036ca19c1ff3dd2ac46a853eb1