[roll] Roll fuchsia [fidlc] Remove derived names
Derived names represent names that correspond to part of a source
file (i.e. have a non-virtual SourceSpan), but that have a separately
specifed name.

This can be considered a more general form of anonymous names, where
the source span refers to the definition of the anonymous object,
and the manually specified name is the flattened name that is reserved
by fidlc.

Since there are no derived names that are not anonymous names, the
code for derived names is removed; things are not renamed to from
"anonymous" to "derived" since "anonymous" is considered to be
more explicit about the use case.

Test: fx ninja -C out/default host_x64/fidl-compiler &&
./out/default/host_x64/fidl-compiler --gtest_filter='TypesTests.*'

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/546662
Original-Revision: 3200302749bee7ad368101dad9e6414b95063cfe
GitOrigin-RevId: 4531ca4a08dc6469184b95f5a5ac53a56b8c21ae
Change-Id: I2aa51f5fdc2dfaa3c5f03a25be7074a0a2247070