[roll] Roll fuchsia [clock][testrunner] Enforce outer test timeout.
Set a slightly longer outer timeout for running a single test so that
the tester can handle the test timeout in its own way, but if it fails
to, then we can fall back to the outer timeout.

We used to use a context timeout to do this, but it was removed in
commit 2315fc48e07c6b861190e6029148410bf9999b5e because it was hitting
this timeout just before it could hit the tester timeout and let the
tester do any necessary cleanup. However, there is still a possibility
that the target's timeout mechanism (which we currently rely on) can
fail, so we should have a larger backup timeout as well.

Original-Bug: 84445
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/580281
Original-Revision: a36dc829c7eca1c1be19e0a25335bda6c70fd255
GitOrigin-RevId: f4475f09f95dc5fdea85591290dfa5875d39c535
Change-Id: I939cb8b2a80717e22178b62c6924c2ff51a9a8c8