[roll] Roll fuchsia [cm][routing] Add optional route mapping to capability router
The currently-used version of the capability route analyzer
records a summary (essentially a list of `{Use,Offer,Expose}Decl`s)
of each route that it walks. This analyzer is not integrated
with the CF routing code -- it has its own implementation of
the routing algorithm, and maintains its own state as it walks
routes. The route summaries are used by some downstream
analyzers: see https://fxrev.dev/568144 for example.

I'm finishing up a new version of the analyzer that *is*
integrated with the CF routing code. In order to swap in the
new analyzer, we need be able to retrieve a summary of the
route after every successful call to `route_capability` from
the static routing analyzer.

The routing algorithm is stateless for most capability types.
To keep that structure, it's preferable for the route mapping
to be a no-op at runtime. The runtime and static component
models are already implemented by separate types, so this
change associates a route mapper type to their shared trait.

Route mapping is not yet complete for capabilities sourced from
collections or from other capabilities.

Original-Bug: 61861

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/579221
Original-Revision: ee5241ce6ff126b5de95cc0494c093f94ef8c0a7
GitOrigin-RevId: a9ac3a77b124ea5d3530da976b800022bac6d52e
Change-Id: I9db9b7529e78d1ba060ffd74c3d678272c2bd16a