[roll] Roll fuchsia [usb-hub] Disconnect all devices on transaction failure.
Previously, when an entire USB hub was disconnected the hub device would
be correctly destroyed, but it would not tear down all of its child
devices, and it would print out an error message "Could not get port
status".

The error message was because the interrupt callback didn't correctly
propogate the response status up to the interrupt thread, which would
cause the thread to sometimes try handling an interrupt that hadn't
happened.

To properly disconnect all USB devices, we disconnect any port that was
previously in the "connected" state after the interrupt loop has
finished (either because unbind() was called or some other error
occurred, e.g. on disconnect).

Multiply: usb-hub-package
Original-Bug: 78415
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/579761
Original-Revision: fce6d27811e0130dc3c63fec5a18682650809612
GitOrigin-RevId: dab5153d7b2a90893953ee0ebf5d8ce3406713e1
Change-Id: I8847e8ff59c3081deb958485e9a04a0bfe264378