[roll] Roll fuchsia [overnet] Map Error::Done to Ok(0)
Due to this change in quiche
https://github.com/cloudflare/quiche/commit/95ea1361794e533bdb75599246f2f92033720621
the future implementation of the quic protocol will return
Err(Error::Done) in some cases where it used to return Ok(0). Overnet
relies on the Ok variant in some cases so upgrading to a quiche with
that change would cause some issues. This change turns that specific
error into the Ok(0) case which overnet can handle. This works with the
current version of quiche and will allow a future upgrade to a version
which includes that change.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/580652
Original-Revision: 344badbd34d9bca745318f89971432a3a10422b1
GitOrigin-RevId: cf55177164f13f8b83f799aba0d47441dbfc939d
Change-Id: I571c6fda7a7886d723177fb077ff2fc271573057