[roll] Roll infra/recipes [buildbucket_util] Fix broken test on Mac
An emoji is represented by two separate 2-byte unicode characters on Mac
compared to one 4-byte character on Linux, leading to inconsistent
results when truncating a string of emojis while avoiding splitting in
the middle of characters. To avoid this issue, just use a more standard
unicode character for testing that doesn't exhibit the same incosistency
as emojis.

There is an underlying bug in our handling of emojis on Mac, but it's
fine to just ignore that for now since it's very much an edge case, and
it will go away once we only support Python 3 anyway.

Original-Bug: 83858
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/580761
Original-Revision: b090630a4625d7e337ca7105f74e9a972bcd0c69
GitOrigin-RevId: 21e22f230ca43e3ae6d3f8d1782354280908c39a
Change-Id: I61f96934b127531d85d78674309b38f6fb11cb0b